### 2025-03-15 #7901 @editorialbot
[PRE REVIEW]: DeepHyper: Massively Parallel Hyperparameter Optimization for Machine Learning
**Submitting author:** @Deathn0t (<a href="http://orcid.org/0000-0002-8992-8192">Romain Egele</a>)
**Repository:** https://github.com/deephyper/deephyper
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.10.0
