+++ examples/examples_hpo/plot_hpo_text_classification.py
@@ -66,6 +66,11 @@ def load_data(train_ratio):
     num_train = int(len(train_dataset) * train_ratio)
     split_train, split_valid = \
         random_split(train_dataset, [num_train, len(train_dataset) - num_train])
+    
+    ## downsample
+    split_train, _ = random_split(split_train, [int(len(split_train)*.1), int(len(split_train)*.9)])
+    split_valid, _ = random_split(split_valid, [int(len(split_valid)*.1), int(len(split_valid)*.9)])
+    test_dataset, _ = random_split(test_dataset, [int(len(test_dataset)*.1), int(len(test_dataset)*.9)])
 
     return split_train, split_valid, test_dataset
 
@@ -138,7 +143,7 @@ class TextClassificationModel(nn.Module):
 
     def __init__(self, vocab_size, embed_dim, num_class):
         super(TextClassificationModel, self).__init__()
-        self.embedding = nn.EmbeddingBag(vocab_size, embed_dim, sparse=True)
+        self.embedding = nn.EmbeddingBag(vocab_size, embed_dim, sparse=False)
         self.fc = nn.Linear(embed_dim, num_class)
         self.init_weights()
 
@@ -218,7 +223,7 @@ def get_run(train_ratio=0.95):
   return run
 
 # %%
-# We create two versions of :code:`run`, one quicker to evaluate for the seacrh, with a small training dataset, and another one, for performance evaluation, which uses a normal training/validation ratio.
+# We create two versions of :code:`run`, one quicker to evaluate for the search, with a small training dataset, and another one, for performance evaluation, which uses a normal training/validation ratio.
 
 # %%
 quick_run = get_run(train_ratio=0.3)
@@ -383,12 +388,10 @@ best_config = {k[2:]: v for k, v in best_config.items() if k.startswith("p:")}
 
 print(f"The default configuration has an accuracy of {objective_default:.3f}. \n" 
       f"The best configuration found by DeepHyper has an accuracy {results['objective'].iloc[i_max]:.3f}, \n" 
-      f"finished after {results['timestamp_gather'].iloc[i_max]:.2f} secondes of search.\n")
+      f"finished after {results['m:timestamp_gather'].iloc[i_max]:.2f} secondes of search.\n")
 
 print(json.dumps(best_config, indent=4))
 
 # %%
 objective_best = perf_run(best_config)
 print(f"Accuracy Best Configuration:  {objective_best:.3f}")
-
-
