+++ docs/examples/examples_hpo/index.rst
@@ -29,6 +29,11 @@ Hyperparameter optimization
 .. raw:: html
 
       <div class="sphx-glr-thumbnail-title">Hyperparameter search for text classification</div>
+    </div>
+
+
+.. raw:: html
+
     <div class="sphx-glr-thumbcontainer" tooltip="In this example, you will learn how to treat the choice of a learning method as just another hyperparameter. We consider the Random Forest (RF) and Gradient Boosting (GB) classifiers from Scikit-Learn on the Airlines dataset.">
 
 .. only:: html
