+++ src/deephyper/skopt/optimizer/optimizer.py
@@ -53,6 +53,7 @@ OBJECTIVE_VALUE_FAILURE = "F"
 
 # TODO: add check to notify that optimizer="ga" cannot work with categorical or integer values
 
+
 class Optimizer(object):
     """Run bayesian optimisation loop in DeepHyper.
 
@@ -278,15 +279,13 @@ class Optimizer(object):
         # Check `n_random_starts` deprecation first
         if n_random_starts is not None:
             warnings.warn(
-                ("n_random_starts will be removed in favour of " "n_initial_points."),
+                ("n_random_starts will be removed in favour of n_initial_points."),
                 DeprecationWarning,
             )
             n_initial_points = n_random_starts
 
         if n_initial_points < 0:
-            raise ValueError(
-                "Expected `n_initial_points` >= 0, got %d" % n_initial_points
-            )
+            raise ValueError("Expected `n_initial_points` >= 0, got %d" % n_initial_points)
         self._n_initial_points = n_initial_points
         self.n_initial_points_ = n_initial_points
 
@@ -314,9 +313,7 @@ class Optimizer(object):
 
         # preprocessing of target variable
         self.objective_scaler_ = objective_scaler
-        self.objective_scaler = cook_objective_scaler(
-            objective_scaler, self.base_estimator_
-        )
+        self.objective_scaler = cook_objective_scaler(objective_scaler, self.base_estimator_)
 
         # Configure optimizer
 
@@ -334,8 +331,9 @@ class Optimizer(object):
             "ga",
         ]:
             raise ValueError(
-                "Expected acq_optimizer to be 'lbfgs' or "
-                "'sampling' or 'ga', got {0}".format(acq_optimizer)
+                "Expected acq_optimizer to be 'lbfgs' or 'sampling' or 'ga', got {0}".format(
+                    acq_optimizer
+                )
             )
 
         self.acq_optimizer = acq_optimizer
@@ -359,9 +357,7 @@ class Optimizer(object):
 
         # TODO: "update_prior" to be removed, this mechanism is too prone to "overfitting" (local minima problems)
         self.update_prior = acq_optimizer_kwargs.get("update_prior", False)
-        self.update_prior_quantile = acq_optimizer_kwargs.get(
-            "update_prior_quantile", 0.9
-        )
+        self.update_prior_quantile = acq_optimizer_kwargs.get("update_prior_quantile", 0.9)
         self.filter_duplicated = acq_optimizer_kwargs.get("filter_duplicated", True)
         self.filter_failures = acq_optimizer_kwargs.get("filter_failures", "mean")
         self.max_failures = acq_optimizer_kwargs.get("max_failures", 100)
@@ -391,18 +387,13 @@ class Optimizer(object):
             self.space.config_space.seed(self.rng.get_state()[1][0])
 
         self._initial_samples = [] if initial_points is None else initial_points[:]
-        self._initial_point_generator = cook_initial_point_generator(
-            initial_point_generator
-        )
+        self._initial_point_generator = cook_initial_point_generator(initial_point_generator)
 
         if self._initial_point_generator is not None:
-            self._initial_samples = (
-                self._initial_samples
-                + self._initial_point_generator.generate(
-                    self.space.dimensions,
-                    n_initial_points - len(self._initial_samples),
-                    random_state=self.rng.randint(0, np.iinfo(np.int32).max),
-                )
+            self._initial_samples = self._initial_samples + self._initial_point_generator.generate(
+                self.space.dimensions,
+                n_initial_points - len(self._initial_samples),
+                random_state=self.rng.randint(0, np.iinfo(np.int32).max),
             )
 
         # record categorical and non-categorical indices
@@ -417,9 +408,7 @@ class Optimizer(object):
[…]
