+++ tests/hpo/test__cbo.py
@@ -680,106 +680,6 @@ def test_cbo_multi_point_strategy(tmp_path):
     assert all(durations[i] > durations[j] for i in range(3) for j in range(3, 5))
 
 
-@pytest.mark.slow
-def test_cbo_with_acq_optimizer_mixedga_and_conditions_in_problem(tmp_path):
-    from ConfigSpace import GreaterThanCondition
-    from deephyper.hpo import CBO, HpProblem
-
-    problem = HpProblem()
-
-    max_num_layers = 3
-    num_layers = problem.add_hyperparameter((1, max_num_layers), "num_layers", default_value=2)
-
-    conditions = []
-    for i in range(max_num_layers):
-        layer_i_units = problem.add_hyperparameter((1, 50), f"layer_{i}_units", default_value=32)
-
-        if i > 0:
-            conditions.extend(
-                [
-                    GreaterThanCondition(layer_i_units, num_layers, i),
-                ]
-            )
-    problem.add_conditions(conditions)
-
-    def run(job):
-        num_layers = job.parameters["num_layers"]
-        return sum(job.parameters[f"layer_{i}_units"] for i in range(num_layers))
-
-    search = CBO(
-        problem,
-        run,
-        n_points=SEARCH_KWARGS_DEFAULTS["n_points"],
-        random_state=SEARCH_KWARGS_DEFAULTS["random_state"],
-        surrogate_model=SEARCH_KWARGS_DEFAULTS["surrogate_model"],
-        surrogate_model_kwargs=SEARCH_KWARGS_DEFAULTS["surrogate_model_kwargs"],
-        log_dir=tmp_path,
-        acq_optimizer="mixedga",
-        acq_optimizer_freq=1,
-        kappa=5.0,
-        scheduler={"type": "periodic-exp-decay", "period": 25, "kappa_final": 0.0001},
-        verbose=0,
-    )
-    results = search.search(max_evals=25)
-
-    assert (results[(results["p:num_layers"] == 1)]["p:layer_1_units"] == 1).all()
-    assert (results[(results["p:num_layers"] == 1)]["p:layer_2_units"] == 1).all()
-    assert (results[(results["p:num_layers"] == 2)]["p:layer_2_units"] == 1).all()
-    assert results["objective"].max() > 100
-
-
-@pytest.mark.slow
-def test_cbo_with_acq_optimizer_mixedga_and_forbiddens_in_problem(tmp_path):
-    from ConfigSpace import ForbiddenAndConjunction, ForbiddenEqualsClause, ForbiddenEqualsRelation
-    from deephyper.hpo import CBO, HpProblem
-
-    problem = HpProblem()
-
-    max_num_layers = 5
-    for i in range(max_num_layers):
-        problem.add_hyperparameter(
-            (0, max_num_layers), f"layer_{i}_units", default_value=(i + 1) % max_num_layers
-        )
-    forbiddens = []
-    for i in range(1, max_num_layers):
-        forb = ForbiddenEqualsRelation(problem[f"layer_{i - 1}_units"], problem[f"layer_{i}_units"])
-        forbiddens.append(forb)
-    problem.add_forbidden_clause(forbiddens)
-    problem.add_forbidden_clause(
-        ForbiddenAndConjunction(
-            *(ForbiddenEqualsClause(problem[f"layer_{i}_units"], i) for i in range(max_num_layers))
-        )
-    )
-    print(problem)
-
-    def run(job):
-        return sum(job.parameters[f"layer_{i}_units"] for i in range(max_num_layers))
-
-    search = CBO(
-        problem,
-        run,
[…]
