### 2025-03-17 adab4d1
refactored mixedga for acq_opt
