+++ src/deephyper/skopt/optimizer/optimizer.py
@@ -1,3 +1,5 @@
+"""Sickit-Optimize optimizer class."""
+
 import numbers
 import sys
 import warnings
@@ -30,20 +32,30 @@ from ..utils import (
 
 
 class ExhaustedSearchSpace(RuntimeError):
-    """ "Raised when the search cannot sample new points from the ConfigSpace."""
+    """Raised when the search cannot sample new points from the ConfigSpace."""
 
-    def __str__(self):
+    def __str__(self):  # noqa: D105
         return "The search space is exhausted and cannot sample new unique points!"
 
 
 class ExhaustedFailures(RuntimeError):
-    """Raised when the search has seen ``max_failures`` failures without any valid objective value."""
+    """Raised when search has seen ``max_failures`` failures without any valid objective value."""
 
-    def __str__(self):
-        return "The search has reached its quota of failures! Check if the type of failure is expected or the value of ``max_failures`` in the search algorithm."
+    def __str__(self):  # noqa: D105
+        out = (
+            "The search has reached its quota of failures!"
+            "Check if the type of failure is expected or the value of ``max_failures`` "
+            "in the search algorithm."
+        )
+        return out
 
 
 def boltzmann_distribution(x, beta=1):
+    """Boltzmann distribution function.
+
+    It is a softmax function with a parameter beta that controls the entropy
+    over probabilities.
+    """
     x = np.exp(beta * x)
     x = x / np.sum(x)
     return x
@@ -1104,86 +1116,42 @@ class Optimizer(object):
                     next_x = cand_xs[np.argmin(cand_acqs)]
 
                 elif self.acq_optimizer == "mixedga":
-                    # TODO: vectorized differential evolution
-                    # https://pymoo.org/customization/mixed.html
-                    # https://pymoo.org/interface/problem.html
-
-                    from pymoo.core.mixed import (
-                        MixedVariableGA,
-                        MixedVariableDuplicateElimination,
-                        MixedVariableMating,
-                    )
-                    from pymoo.core.population import Population
-                    from pymoo.optimize import minimize
-                    from pymoo.termination.default import (
-                        DefaultSingleObjectiveTermination,
+                    from deephyper.skopt.optimizer.acq_optimizer.pymoo_mixedga import (
+                        MixedGAPymooAcqOptimizer,
                     )
 
-                    from deephyper.skopt.optimizer._pymoo import (
-                        DefaultSingleObjectiveMixedTermination,
-                        PyMOOMixedVectorizedProblem,
-                        ConfigSpaceRepair,
-                    )
+                    pop_size = self._pymoo_pop_size
 
-                    pop = self._pymoo_pop_size
                     idx_sorted = np.argsort(values)
-                    initial_sampling = [Xsample[i] for i in idx_sorted[:pop]]
-                    initial_sampling = list(
+                    x_init = [Xsample[i] for i in idx_sorted[:pop_size]]
+                    x_init = list(
                         map(
                             lambda x: dict(zip(self.space.dimension_names, x)),
-                            initial_sampling,
+                            x_init,
                         )
                     )
-                    init_pop = Population.new(
-                        "X",
-                        initial_sampling,
-                        "F",
-                        values[idx_sorted[:pop]].reshape(-1),
+
+                    acq_opt = MixedGAPymooAcqOptimizer(
+                        space=self.space,
+                        x_init=x_init,
+                        y_init=values[idx_sorted[:pop_size]],
+                        pop_size=self._pymoo_pop_size,
+                        random_state=self.rng.randint(0, np.iinfo(np.int32).max),
+                        termination_kwargs=self._pymoo_termination_kwargs,
                     )
 
                     args = (est, np.min(yi), cand_acq_func, False, self.acq_func_kwargs)
[…]
