### 2025-03-17 2589789
updated skopt optimizer with ruff formatting
