### 2025-03-17 #7901 @jbytecode
Dear @Deathn0t,

Thank you for submitting your work to JOSS. I'm the handling editor.

In your paper.md, it seems the citations are declared manually, e.g., `(Egele et al., 2022)`, but they should be hosted in the bibliography file and be referenced using the markdown notation, please see

https://joss.readthedocs.io/en/latest/paper.html#citations

and fix the paper as the citations are properly referenced. Please ping me when you've done with it.

The next step is to find at least two suitable reviewers. Do you have any suggestions for potential reviewers for the submission? You can use the suggestions of our editorialbot and filter their authors if they are suitable (please take a look at the post above). Addition to this, you can use the list of people, see:

https://reviewers.joss.theoj.org/lookup

Thank you in advance!
