### 2025-03-17 860fa3f
updating citations in joss paper.md with direct [@...] reference
