### 2025-03-17 fa5be42
adding paragraph on GMM
