### 2025-03-17 9750bc0
adding missing dois to paper.bib for joss
