+++ paper.bib
@@ -2,6 +2,7 @@
   title        = {DeepHyper: Asynchronous hyperparameter search for deep neural networks},
   author       = {Balaprakash, Prasanna and Salim, Michael and Uram, Thomas D. and Vishwanath, Venkat and Wild, Stefan M.},
   booktitle    = {IEEE 25th international conference on high performance computing (HiPC)},
+  doi          = {10.1109/HiPC.2018.00014},
   pages        = {42--51},
   year         = {2018},
   organization = {IEEE}
@@ -29,6 +30,7 @@
   title     = {Recurrent Neural Network Architecture Search for Geophysical Emulation},
   year      = {2020},
   isbn      = {9781728199986},
+  doi       = {10.1109/SC41405.2020.00012},
   publisher = {IEEE Press},
   booktitle = {Proceedings of the International Conference for High Performance Computing, Networking, Storage and Analysis},
   articleno = {8},
@@ -59,6 +61,7 @@
   title        = {{AutoDEUQ}: Automated deep ensemble with uncertainty quantification},
   author       = {Égelé, Romain and Maulik, Romit and Raghavan, Krishnan and Lusch, Bethany and Guyon, Isabelle and Balaprakash, Prasanna},
   booktitle    = {26th International Conference on Pattern Recognition (ICPR)},
+  doi          = {10.1109/ICPR56361.2022.9956231},
   pages        = {1908--1914},
   year         = {2022},
   organization = {IEEE}
@@ -84,6 +87,7 @@
   author       = {Egel{\'e}, Romain and Guyon, Isabelle and Vishwanath, Venkatram and Balaprakash, Prasanna},
   booktitle    = {2023 IEEE 19th International Conference on e-Science (e-Science)},
   pages        = {1--10},
+  doi          = {10.1109/e-Science58273.2023.10254839},
   year         = {2023},
   organization = {IEEE}
 }
@@ -120,10 +124,11 @@
 }
 
 @article{egele2024unreasonable,
-  title={The unreasonable effectiveness of early discarding after one epoch in neural network hyperparameter optimization},
-  author={Egele, Romain and Mohr, Felix and Viering, Tom and Balaprakash, Prasanna},
-  journal={Neurocomputing},
-  pages={127964},
-  year={2024},
-  publisher={Elsevier}
+  title     = {The unreasonable effectiveness of early discarding after one epoch in neural network hyperparameter optimization},
+  author    ={Egele, Romain and Mohr, Felix and Viering, Tom and Balaprakash, Prasanna},
+  journal   = {Neurocomputing},
+  pages     = {127964},
+  year      = {2024},
+  doi       = {10.1016/j.neucom.2024.127964},
+  publisher = {Elsevier}
 }
