### 2025-03-17 0adc950
removed "update_prior" from cbo and skopt space
