+++ src/deephyper/hpo/_cbo.py
@@ -158,13 +158,6 @@ class CBO(Search):
             search space is "exhausted" in the sens that no new unique points can be found given
             the sampling size ``n_points``. Defaults to ``True``.
 
-        update_prior (bool, optional): Update the prior of the surrogate model with the new
-            evaluated points. Defaults to ``False``. Should be set to ``True`` when all objectives
-            and parameters are continuous.
-
-        update_prior_quantile (float, optional): The quantile used to update the prior.
-            Defaults to ``0.1``.
-
         multi_point_strategy (str, optional): Definition of the constant value use for the Liar
             strategy. Can be a value in ``["cl_min", "cl_mean", "cl_max", "qUCB", "qUCBd"]``. All
             ``"cl_..."`` strategies follow the constant-liar scheme, where if $N$ new points are
@@ -252,8 +245,6 @@ class CBO(Search):
         xi: float = 0.001,
         n_points: int = 10_000,
         filter_duplicated: bool = True,
-        update_prior: bool = False,
-        update_prior_quantile: float = 0.1,
         multi_point_strategy: str = "cl_max",
         n_jobs: int = 1,  # 32 is good for Theta
         n_initial_points: int = 10,
@@ -416,8 +407,6 @@ class CBO(Search):
             acq_optimizer_kwargs={
                 "n_points": n_points,
                 "filter_duplicated": filter_duplicated,
-                "update_prior": update_prior,
-                "update_prior_quantile": 1 - update_prior_quantile,
                 "n_jobs": n_jobs,
                 "filter_failures": MAP_filter_failures.get(filter_failures, filter_failures),
                 "max_failures": max_failures,
