### 2025-03-17 #7901 @jbytecode
👋👋👋 Dear @maxpumperla,  @jungtaekkim, @jmlarson1, @hawk31,  @geoffreyp 👋👋👋

**[DeepHyper: Massively Parallel Hyperparameter Optimization for Machine Learning](https://github.com/openjournals/joss-reviews/issues/7901)**

You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/7901)).

Here is the reviewing guidelines for reviewing in JOSS: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Thank you in advance!
