### 2025-03-18 9973089
convert and move tutorial to example in new format
