### 2025-03-18 10ca297
updated stopping criteria for acq_opt with pymoo
