+++ src/deephyper/skopt/optimizer/acq_optimizer/pymoo_ga.py
@@ -64,10 +64,10 @@ class GAPymooAcqOptimizer:
         if termination_kwargs is None:
             termination_kwargs = {}
         default_termination_kwargs = {
-            "xtol": 1e-3,
-            "ftol": 1e-3,
-            "period": 15,
-            "n_max_gen": 1000,
+            "xtol": 1e-8,
+            "ftol": 1e-6,
+            "period": 30,
+            "n_max_gen": 1_000,
         }
         default_termination_kwargs.update(termination_kwargs)
         self.termination_kwargs = default_termination_kwargs
