### 2025-03-18 4c4fd6d
Convert tutorial on HPO for text-classification with PyTorch to example in sphinx-gallery(#298)
Co-authored-by: Brett Eiffert <eiffertbc@ornl.gov>
