+++ docs/examples/examples_hpo/sg_execution_times.rst
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**02:45.186** total execution time for 1 file **from examples/examples_hpo**:
+**07:12.166** total execution time for 2 files **from examples/examples_hpo**:
 
 .. container::
 
@@ -32,6 +32,9 @@ Computation times
    * - Example
      - Time
      - Mem (MB)
+   * - :ref:`sphx_glr_examples_examples_hpo_plot_hpo_text_classification.py` (``plot_hpo_text_classification.py``)
+     - 07:12.166
+     - 0.0
    * - :ref:`sphx_glr_examples_examples_hpo_plot_hpo_for_rf_and_overfitting.py` (``plot_hpo_for_rf_and_overfitting.py``)
-     - 02:45.186
+     - 00:00.000
      - 0.0
