### 2025-03-21 34f9772
adding pydantic checks for cbo parameters, adding scheduler for surrogate model parameters in cbo
