+++ src/deephyper/skopt/optimizer/optimizer.py
@@ -66,6 +66,11 @@ OBJECTIVE_VALUE_FAILURE = "F"
 # TODO: add check to notify that optimizer="ga" cannot work with categorical or integer values
 
 
+def is_not_improving(scores, patience):
+    """Check if scores have improved over 'patience' length."""
+    return len(scores) > (patience + 1) and scores[-patience - 1] < min(scores[-patience:])
+
+
 class Optimizer(object):
     r"""Run bayesian optimisation loop in DeepHyper.
 
@@ -190,7 +195,7 @@ class Optimizer(object):
         custom_sampler (Model or None, optional): Default None
             A Model from Synthetic-Data-Vault.
 
-        moo_scalarization_strategy (string, optional): Default is `"Chebyshev"`
+        moo_scalarization_strategy (string, optional): Default is ``"Chebyshev"``.
             Function to convert multiple objectives into a single scalar value. Can be either
 
             - `"Linear"` for linear/convex combination.
@@ -222,20 +227,19 @@ class Optimizer(object):
         self,
         dimensions,
         base_estimator="gp",
+        base_estimator_scheduler=None,
         n_random_starts=None,
         n_initial_points=10,
         initial_points=None,
         initial_point_generator="random",
         n_jobs=1,
         acq_func="gp_hedge",
+        acq_func_kwargs=None,
         acq_optimizer="auto",
+        acq_optimizer_kwargs=None,
         random_state=None,
         model_queue_size=None,
-        acq_func_kwargs=None,
-        acq_optimizer_kwargs=None,
         custom_sampler=None,
-        sample_max_size=-1,
-        sample_strategy="quantile",
         moo_upper_bounds=None,
         moo_scalarization_strategy="Chebyshev",
         moo_scalarization_weight=None,
@@ -324,6 +328,12 @@ class Optimizer(object):
         else:
             self.base_estimator_ = base_estimator
 
+        self.base_estimator_scheduler = base_estimator_scheduler
+        if self.base_estimator_scheduler is not None and not isinstance(
+            self.base_estimator_scheduler, dict
+        ):
+            raise ValueError("base_estimator_schedule should be a dict")
+
         # preprocessing of target variable
         self.objective_scaler_ = objective_scaler
         self.objective_scaler = cook_objective_scaler(objective_scaler, self.base_estimator_)
@@ -360,12 +370,12 @@ class Optimizer(object):
         self.n_jobs = acq_optimizer_kwargs.get("n_jobs", 1)
 
         # PyMOO
-        self._pymoo_pop_size = acq_optimizer_kwargs.get("pop_size", 100)
+        self._pymoo_pop_size = acq_optimizer_kwargs.get("ga_pop_size", 100)
         self._pymoo_termination_kwargs = {
-            "xtol": acq_optimizer_kwargs.get("xtol", 1e-8),
-            "ftol": acq_optimizer_kwargs.get("ftol", 1e-6),
-            "period": acq_optimizer_kwargs.get("period", 30),
-            "n_max_gen": acq_optimizer_kwargs.get("n_max_gen", 1000),
+            "xtol": acq_optimizer_kwargs.get("ga_xtol", 1e-8),
+            "ftol": acq_optimizer_kwargs.get("ga_ftol", 1e-6),
+            "period": acq_optimizer_kwargs.get("ga_period", 30),
+            "n_max_gen": acq_optimizer_kwargs.get("ga_n_max_gen", 1000),
         }
 
         self.filter_duplicated = acq_optimizer_kwargs.get("filter_duplicated", True)
@@ -470,10 +480,6 @@ class Optimizer(object):
         self._min_value = 0
         self._max_value = 0
 
[…]
