+++ src/deephyper/hpo/_cbo.py
@@ -3,7 +3,8 @@ import logging
 import numbers
 import time
 import warnings
-from typing import Dict, List
+from typing import Dict, List, Optional, Literal, Union
+from pydantic import BaseModel, field_validator, ValidationInfo
 
 import ConfigSpace as CS
 import ConfigSpace.hyperparameters as csh
@@ -37,6 +38,54 @@ MAP_multi_point_strategy = {
 MAP_acq_func = {"UCB": "LCB", "UCBd": "LCBd"}
 
 MAP_filter_failures = {"min": "max"}
+
+
+class AcqFuncKwargsScheduler(BaseModel):
+    type: Optional[Literal["bandit", "periodic-exp-decay"]] = "periodic-exp-decay"
+    delay: Optional[Union[int, Literal["n-initial-points"]]] = "n-initial-points"
+    # "periodic-exp-decay" parameters
+    kappa_final: Optional[float] = 0.01
+    xi_final: Optional[float] = 1e-4
+    period: Optional[int] = 10
+    # "bandit" parameters
+    delta: Optional[float] = 0.05
+    lamb: Optional[float] = 0.2
+
+
+class AcqFuncKwargs(BaseModel):
+    kappa: Optional[float] = 1.96
+    xi: Optional[float] = 0.1
+    scheduler: Optional[AcqFuncKwargsScheduler] = dict(AcqFuncKwargsScheduler())
+
+
+class AcqOptimizerKwargs(BaseModel):
+    n_points: Optional[int] = 10_000
+    filter_duplicated: Optional[bool] = True
+    filter_failures: Optional[Literal["ignore", "max", "mean"]] = "max"
+    max_failures: Optional[int] = 100
+    acq_optimizer_freq: Optional[int] = 1
+    n_jobs: Optional[int] = 1
+    n_restarts_optimizer: Optional[int] = 1
+    # Genetic algorithms parameters
+    ga_pop_size: Optional[int] = 100
+    ga_xtol: Optional[float] = 1e-8
+    ga_ftol: Optional[float] = 1e-6
+    ga_period: Optional[int] = 30
+    ga_n_max_gen: Optional[int] = 1000
+
+    @field_validator("filter_failures")
+    @classmethod
+    def map_filter_failures(cls, v: str, info: ValidationInfo) -> str:
+        return MAP_filter_failures.get(v, v)
+
+
+class SurrogateModelScheduler(BaseModel):
+    patience: Optional[int] = 10
+    params: Optional[dict] = dict(min_impurity_decrease=dict(factor=0.2))
+
+
+class SurrogateModelKwargs(BaseModel):
+    scheduler: Optional[SurrogateModelScheduler] = dict(SurrogateModelScheduler())
 
 
 # schedulers
@@ -138,25 +187,52 @@ class CBO(Search):
         acq_func (str, optional): Acquisition function used by the Bayesian optimization. Can be a
             value in ``["UCB", "EI", "PI", "gp_hedge"]``. Defaults to ``"UCB"``.
 
-        acq_optimizer (str, optional): Method used to minimze the acquisition function. Can be a
-            value in ``["sampling", "lbfgs", "ga", "mixedga"]``. Defaults to ``"auto"``.
+        acq_func_kwargs (dict, optional):
+            A dictionnary of parameters for the acquisition function:
 
-        acq_optimizer_freq (int, optional): Frequency of optimization calls for the acquisition
-            function. Defaults to ``10``, using optimizer every ``10`` surrogate model updates.
+            - ``"kappa"`` (float)
+                Manage the exploration/exploitation tradeoff for the ``"UCB"`` acquisition function.
+                Defaults to ``1.96`` which corresponds to 95% of the confidence interval.
 
[…]
