+++ tests/hpo/test__cbo.py
@@ -8,6 +8,7 @@ SEARCH_KWARGS_DEFAULTS = dict(
     surrogate_model="ET",
     surrogate_model_kwargs={"n_estimators": 25, "min_samples_split": 8},
     acq_func_kwargs={"kappa": "toto"},
+    acq_optimizer="sampling",
     acq_optimizer_kwargs={"n_points": 100},
 )
 
@@ -28,6 +29,7 @@ def test_cbo_random_seed(tmp_path):
     search = CBO(
         problem,
         create_evaluator(),
+        acq_optimizer=SEARCH_KWARGS_DEFAULTS["acq_optimizer"],
         acq_optimizer_kwargs=SEARCH_KWARGS_DEFAULTS["acq_optimizer_kwargs"],
         random_state=SEARCH_KWARGS_DEFAULTS["random_state"],
         surrogate_model="DUMMY",
@@ -40,6 +42,7 @@ def test_cbo_random_seed(tmp_path):
     search = CBO(
         problem,
         create_evaluator(),
+        acq_optimizer=SEARCH_KWARGS_DEFAULTS["acq_optimizer"],
         acq_optimizer_kwargs=SEARCH_KWARGS_DEFAULTS["acq_optimizer_kwargs"],
         random_state=SEARCH_KWARGS_DEFAULTS["random_state"],
         surrogate_model="DUMMY",
@@ -60,6 +63,7 @@ def test_cbo_random_seed(tmp_path):
     search = CBO(
         problem,
         create_evaluator(),
+        acq_optimizer=SEARCH_KWARGS_DEFAULTS["acq_optimizer"],
         acq_optimizer_kwargs=SEARCH_KWARGS_DEFAULTS["acq_optimizer_kwargs"],
         random_state=SEARCH_KWARGS_DEFAULTS["random_state"],
         surrogate_model="DUMMY",
@@ -72,6 +76,7 @@ def test_cbo_random_seed(tmp_path):
     search = CBO(
         problem,
         create_evaluator(),
+        acq_optimizer=SEARCH_KWARGS_DEFAULTS["acq_optimizer"],
         acq_optimizer_kwargs=SEARCH_KWARGS_DEFAULTS["acq_optimizer_kwargs"],
         random_state=SEARCH_KWARGS_DEFAULTS["random_state"],
         surrogate_model="DUMMY",
@@ -103,6 +108,7 @@ def test_sample_types(tmp_path):
         problem,
         run,
         n_initial_points=5,
+        acq_optimizer=SEARCH_KWARGS_DEFAULTS["acq_optimizer"],
         acq_optimizer_kwargs=SEARCH_KWARGS_DEFAULTS["acq_optimizer_kwargs"],
         random_state=SEARCH_KWARGS_DEFAULTS["random_state"],
         surrogate_model="DUMMY",
@@ -115,6 +121,7 @@ def test_sample_types(tmp_path):
         problem,
         run,
         n_initial_points=5,
+        acq_optimizer=SEARCH_KWARGS_DEFAULTS["acq_optimizer"],
         acq_optimizer_kwargs=SEARCH_KWARGS_DEFAULTS["acq_optimizer_kwargs"],
         random_state=SEARCH_KWARGS_DEFAULTS["random_state"],
         surrogate_model=SEARCH_KWARGS_DEFAULTS["surrogate_model"],
@@ -143,6 +150,7 @@ def test_sample_types_no_cat(tmp_path):
         problem,
         run,
         random_state=SEARCH_KWARGS_DEFAULTS["random_state"],
+        acq_optimizer=SEARCH_KWARGS_DEFAULTS["acq_optimizer"],
         acq_optimizer_kwargs=SEARCH_KWARGS_DEFAULTS["acq_optimizer_kwargs"],
         n_initial_points=5,
[…]
