### 2025-03-21 853ab06
changing dict of pydantic object to model_dump
