+++ src/deephyper/hpo/_mpi_dbo.py
@@ -1,5 +1,6 @@
 import asyncio
 import logging
+from typing import Optional
 
 import numpy as np
 import scipy.stats
@@ -8,14 +9,10 @@ from deephyper.evaluator import Evaluator, HPOJob
 from deephyper.evaluator.callback import TqdmCallback
 from deephyper.evaluator.mpi import MPI
 from deephyper.evaluator.storage import Storage
-from deephyper.hpo._cbo import CBO
+from deephyper.hpo._cbo import CBO, AcqFuncKwargs, AcqOptimizerKwargs, SurrogateModelKwargs
 from deephyper.stopper import Stopper
 
 __all__ = ["MPIDistributedBO"]
-
-MAP_acq_func = {
-    "UCB": "LCB",
-}
 
 
 class MPIDistributedBO(CBO):
@@ -39,8 +36,8 @@ class MPIDistributedBO(CBO):
 
         random_state (int, optional): Random seed. Defaults to ``None``.
 
-        log_dir (str, optional): Log directory where search's results are saved.
-            Defaults to ``"."``.
+        log_dir (str, optional): Log directory where search's results are saved. Defaults to
+            ``"."``.
 
         verbose (int, optional): Indicate the verbosity level of the search. Defaults to ``0``.
 
@@ -60,46 +57,72 @@ class MPIDistributedBO(CBO):
             other tree-based method worse uncertainty quantification capabilities and slower than
             ``"RF"``. Defaults to ``"ET"``.
 
+        surrogate_model_kwargs (dict, optional): Additional parameters to pass to the surrogate
+            model. Defaults to ``None``.
+
         acq_func (str, optional): Acquisition function used by the Bayesian optimization. Can be a
             value in ``["UCB", "EI", "PI", "gp_hedge"]``. Defaults to ``"UCB"``.
 
-        acq_optimizer (str, optional): Method used to minimze the acquisition function. Can be a
-            value in ``["sampling", "lbfgs"]``. Defaults to ``"auto"``.
+        acq_func_kwargs (dict, optional):
+            A dictionnary of parameters for the acquisition function:
 
-        acq_optimizer_freq (int, optional): Frequency of optimization calls for the acquisition
-            function. Defaults to ``10``, using optimizer every ``10`` surrogate model updates.
+            - ``"kappa"`` (float)
+                Manage the exploration/exploitation tradeoff for the ``"UCB"`` acquisition function.
+                Defaults to ``1.96`` which corresponds to 95% of the confidence interval.
 
-        kappa (float, optional): Manage the exploration/exploitation tradeoff for the "UCB"
-            acquisition function. Defaults to ``1.96`` which corresponds to 95% of the confidence
-            interval.
+            - ``"xi"`` (float)
+                Manage the exploration/exploitation tradeoff of ``"EI"`` and ``"PI"``
+                acquisition function. Defaults to ``0.001``.
 
-        xi (float, optional): Manage the exploration/exploitation tradeoff of ``"EI"`` and
-            ``"PI"`` acquisition function. Defaults to ``0.001``.
+        acq_optimizer (str, optional):
+            Method used to minimze the acquisition function. Can be a value in
+            ``["sampling", "lbfgs", "ga", "mixedga"]``. Defaults to ``"auto"``.
 
-        n_points (int, optional): The number of configurations sampled from the search space to
-            infer each batch of new evaluated configurations.
+        acq_optimizer_kwargs (dict, optional):
+            A dictionnary of parameters for the acquisition function optimizer:
 
-        filter_duplicated (bool, optional): Force the optimizer to sample unique points until the
-            search space is "exhausted" in the sens that no new unique points can be found given
-            the sampling size ``n_points``. Defaults to ``True``.
[…]
