### 2025-03-26 b572824
surrogate_model_kwargs now accepts extra params
