### 2025-03-26 b0f1b79
updating examples
