+++ examples/examples_bbo/plot_multi_objective_optimization.py
@@ -29,30 +29,21 @@ We will look at the DTLZ benchmark suite, a classic in multi-objective optimizat
 #     pip install -e "git+https://github.com/deephyper/benchmark.git@main#egg=deephyper-benchmark"
 
 # .. dropdown:: Import statements
-import os
-
 import matplotlib.pyplot as plt
+
+from deephyper.hpo import CBO
+from deephyper_benchmark.benchmarks.dtlz import DTLZBenchmark
 
 WIDTH_PLOTS = 8
 HEIGHT_PLOTS = WIDTH_PLOTS / 1.618
 
 n_objectives = 2
-
-# Configuration of the DTLZ Benchmark
-os.environ["DEEPHYPER_BENCHMARK_DTLZ_PROB"] = str(2)
-os.environ["DEEPHYPER_BENCHMARK_NDIMS"] = str(8)
-os.environ["DEEPHYPER_BENCHMARK_NOBJS"] = str(n_objectives)
-os.environ["DEEPHYPER_BENCHMARK_DTLZ_OFFSET"] = str(0.6)
-os.environ["DEEPHYPER_BENCHMARK_FAILURES"] = str(0)
-
-# Loading the DTLZ Benchmark
-import deephyper_benchmark as dhb; dhb.load("DTLZ");
-from deephyper_benchmark.lib.dtlz import hpo, metrics
+bench = DTLZBenchmark(nobj=n_objectives)
 
 # %%
 # We can display the variable search space of the benchmark we just loaded:
 
-hpo.problem
+bench.problem
 
 # %%
 # To define a black-box for multi-objective optimization it is very similar to single-objective optimization at the difference that the ``objective`` can now be a list of values. A first possibility is:
@@ -81,49 +72,22 @@ hpo.problem
 # each of the metadata needs to be JSON serializable and will be returned in the final results with a column name formatted as ``m:metadata_key`` such as ``m:duration``.
 
 # %%
-# Now we can load Centralized Bayesian Optimization search:
-
-from deephyper.hpo import CBO
-from deephyper.evaluator import Evaluator
-from deephyper.evaluator.callback import TqdmCallback
-
-# %%
-# Interface to submit/gather parallel evaluations of the black-box function.
-# The method argument is used to specify the parallelization method, in our case we use threads.
-# The method_kwargs argument is used to specify the number of workers and the callbacks.
-# The TqdmCallback is used to display a progress bar during the search.
-
-evaluator = Evaluator.create(
-    hpo.run,
-    method="thread",
-    method_kwargs={"num_workers": 4, "callbacks": [TqdmCallback()]},
-)
-
-# %%
+# For the search algorithm, we use the centralized Bayesian Optimization search (CBO).
 # Search algorithm
-# The acq_func argument is used to specify the acquisition function.
-# The multi_point_strategy argument is used to specify the multi-point strategy,
-# in our case we use qUCB instead of the default cl_max (constant-liar) to reduce overheads.
-# The update_prior argument is used to specify whether the sampling-prior should
-# be updated during the search.
-# The update_prior_quantile argument is used to specify the quantile of the lower-bound
-# used to update the sampling-prior.
-# The moo_scalarization_strategy argument is used to specify the scalarization strategy.
-# Chebyshev is capable of generating a diverse set of solutions for non-convex problems.
-# The moo_scalarization_weight argument is used to specify the weight of the scalarization.
-# random is used to generate a random weight vector for each iteration.
+#
+# The arguments specific to multi-objective optimization are:
+# 
[…]
