+++ docs/examples/examples_bbo/plot_black_box_optimization.rst
@@ -56,7 +56,7 @@ Then, we can import it and check the installed version:
 
  .. code-block:: none
 
-    0.9.2
+    0.9.3
 
 
 
@@ -181,7 +181,7 @@ Search Algorithm
 
 The next step is to define the search algorithm that we want to use. Here, we choose :class:`deephyper.hpo.CBO` (Centralized Bayesian Optimization) which is a sampling based Bayesian optimization strategy. This algorithm has the advantage of being asynchronous thanks to a constant liar strategy which is crutial to keep a good utilization of the resources when the number of available workers increases.
 
-.. GENERATED FROM PYTHON SOURCE LINES 95-107
+.. GENERATED FROM PYTHON SOURCE LINES 95-105
 
 .. code-block:: Python
 
@@ -192,9 +192,7 @@ The next step is to define the search algorithm that we want to use. Here, we ch
     search = CBO(
         problem,
         evaluator,
-        acq_func="UCB",  # Acquisition function to Upper Confidence Bound
-        multi_point_strategy="qUCB",  # Fast Multi-point strategy with q-Upper Confidence Bound
-        n_jobs=2,  # Number of threads to fit surrogate models in parallel
+        acq_optimizer="ga",
     )
 
 
@@ -205,16 +203,16 @@ The next step is to define the search algorithm that we want to use. Here, we ch
 
  .. code-block:: none
 
-    WARNING:root:Results file already exists, it will be renamed to /Users/35e/Projects/DeepHyper/deephyper/examples/examples_bbo/results_20250228-121430.csv
+    WARNING:root:Results file already exists, it will be renamed to /Users/romainegele/Documents/DeepHyper/deephyper/examples/examples_bbo/results_20250324-091742.csv
 
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 108-109
+.. GENERATED FROM PYTHON SOURCE LINES 106-107
 
 Then, we can execute the search for a given number of iterations by using the ``search.search(max_evals=...)``. It is also possible to use the ``timeout`` parameter if one needs a specific time budget (e.g., restricted computational time in machine learning competitions, allocation time in HPC).
 
-.. GENERATED FROM PYTHON SOURCE LINES 111-113
+.. GENERATED FROM PYTHON SOURCE LINES 109-111
 
 .. code-block:: Python
 
@@ -228,11 +226,14 @@ Then, we can execute the search for a given number of iterations by using the ``
 
  .. code-block:: none
 
[…]
