+++ docs/examples/examples_bbo/sg_execution_times.rst
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**02:23.304** total execution time for 5 files **from examples/examples_bbo**:
+**00:21.313** total execution time for 5 files **from examples/examples_bbo**:
 
 .. container::
 
@@ -32,8 +32,8 @@ Computation times
    * - Example
      - Time
      - Mem (MB)
-   * - :ref:`sphx_glr_examples_examples_bbo_plot_notify_failures_hpo.py` (``plot_notify_failures_hpo.py``)
-     - 02:23.304
+   * - :ref:`sphx_glr_examples_examples_bbo_plot_multi_objective_optimization.py` (``plot_multi_objective_optimization.py``)
+     - 00:21.313
      - 0.0
    * - :ref:`sphx_glr_examples_examples_bbo_plot_black_box_optimization.py` (``plot_black_box_optimization.py``)
      - 00:00.000
@@ -41,7 +41,7 @@ Computation times
    * - :ref:`sphx_glr_examples_examples_bbo_plot_experimental_design.py` (``plot_experimental_design.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_examples_examples_bbo_plot_multi_objective_optimization.py` (``plot_multi_objective_optimization.py``)
+   * - :ref:`sphx_glr_examples_examples_bbo_plot_notify_failures_hpo.py` (``plot_notify_failures_hpo.py``)
      - 00:00.000
      - 0.0
    * - :ref:`sphx_glr_examples_examples_bbo_plot_transfer_learning_for_hpo.py` (``plot_transfer_learning_for_hpo.py``)
