### 2025-03-26 783da5c
adding close of tqdm bar in callback
