### 2025-03-26 9cc4361
updating examples
