+++ docs/examples/examples_bbo/plot_notify_failures_hpo.rst
@@ -167,11 +167,11 @@ We will compare:
         search = CBO(
             problem,
             evaluator,
-            acq_func="UCBd",
             acq_optimizer="ga",
-            acq_optimizer_freq=1,
-            filter_duplicated=False,
-            filter_failures=failure_strategy,
+            acq_optimizer_kwargs=dict(
+                filter_duplicated=False,
+                filter_failures=failure_strategy,
+            ),
             log_dir=f"search_{failure_strategy}",
             random_state=42,
         )
@@ -186,168 +186,10 @@ We will compare:
 
  .. code-block:: none
 
-    WARNING:root:Results file already exists, it will be renamed to /Users/romainegele/Documents/DeepHyper/deephyper/examples/examples_bbo/search_ignore/results_20250311-100338.csv
-      0%|          | 0/100 [00:00<?, ?it/s]    Failure Strategy: ignore:   0%|          | 0/100 [00:00<?, ?it/s]    Failure Strategy: ignore:   1%|          | 1/100 [00:00<00:00, 3449.26it/s, failures=0, objective=6.63]    Failure Strategy: ignore:   2%|▏         | 2/100 [00:00<00:00, 232.18it/s, failures=1, objective=6.63]     Failure Strategy: ignore:   3%|▎         | 3/100 [00:00<00:00, 194.92it/s, failures=2, objective=6.63]    Failure Strategy: ignore:   4%|▍         | 4/100 [00:00<00:00, 179.69it/s, failures=2, objective=6.63]    Failure Strategy: ignore:   5%|▌         | 5/100 [00:00<00:00, 172.86it/s, failures=2, objective=6.63]    Failure Strategy: ignore:   6%|▌         | 6/100 [00:00<00:00, 170.10it/s, failures=3, objective=6.63]    Failure Strategy: ignore:   7%|▋         | 7/100 [00:00<00:00, 168.54it/s, failures=4, objective=6.63]    Failure Strategy: ignore:   8%|▊         | 8/100 [00:00<00:00, 166.40it/s, failures=4, objective=6.63]    Failure Strategy: ignore:   9%|▉         | 9/100 [00:00<00:00, 165.35it/s, failures=5, objective=6.63]    Failure Strategy: ignore:  10%|█         | 10/100 [00:00<00:00, 163.99it/s, failures=5, objective=6.63]    Failure Strategy: ignore:  11%|█         | 11/100 [00:00<00:00, 163.54it/s, failures=5, objective=6.63]    Failure Strategy: ignore:  12%|█▏        | 12/100 [00:00<00:00, 163.08it/s, failures=6, objective=6.63]    Failure Strategy: ignore:  13%|█▎        | 13/100 [00:00<00:00, 162.89it/s, failures=6, objective=6.63]    Failure Strategy: ignore:  14%|█▍        | 14/100 [00:00<00:00, 162.55it/s, failures=7, objective=6.63]    Failure Strategy: ignore:  15%|█▌        | 15/100 [00:00<00:00, 162.42it/s, failures=8, objective=6.63]    Failure Strategy: ignore:  16%|█▌        | 16/100 [00:00<00:00, 162.20it/s, failures=9, objective=6.63]    Failure Strategy: ignore:  17%|█▋        | 17/100 [00:00<00:00, 162.14it/s, failures=9, objective=6.63]    Failure Strategy: ignore:  17%|█▋        | 17/100 [00:00<00:00, 162.14it/s, failures=10, objective=6.63]    Failure Strategy: 
[…]
