### 2025-03-26 0f07efb
adding before in field validator of acq_optimizer_kwargs
