### 2025-03-26 eb622c3
updated examples
