+++ docs/examples/examples_parallelism/plot_profile_worker_utilization.rst
@@ -165,7 +165,7 @@ Then we define a parallel search.
  As the ``run``-function is defined in the same module  we use the "loky" backend 
 that serialize by value.
 
-.. GENERATED FROM PYTHON SOURCE LINES 86-112
+.. GENERATED FROM PYTHON SOURCE LINES 86-113
 
 .. code-block:: Python
 
@@ -183,6 +183,7 @@ that serialize by value.
         search = CBO(
             problem,
             evaluator,
+            multi_point_strategy="qUCBd",
             random_state=42,
         )
 
@@ -203,15 +204,15 @@ that serialize by value.
 
  .. code-block:: none
 
-    0it [00:00, ?it/s]    1it [00:00, 4505.16it/s, failures=0, objective=-21.5]    2it [00:00, 12.79it/s, failures=0, objective=-21.5]      2it [00:00, 12.79it/s, failures=0, objective=-19.8]    3it [00:00, 12.79it/s, failures=0, objective=-19.8]    4it [00:00,  9.29it/s, failures=0, objective=-19.8]    4it [00:00,  9.29it/s, failures=0, objective=-19.8]    5it [00:01,  2.22it/s, failures=0, objective=-19.8]    5it [00:01,  2.22it/s, failures=0, objective=-15.4]    6it [00:02,  1.80it/s, failures=0, objective=-15.4]    6it [00:02,  1.80it/s, failures=0, objective=-15.4]    7it [00:02,  2.27it/s, failures=0, objective=-15.4]    7it [00:02,  2.27it/s, failures=0, objective=-15.4]    8it [00:02,  2.27it/s, failures=0, objective=-15.4]    9it [00:04,  1.75it/s, failures=0, objective=-15.4]    9it [00:04,  1.75it/s, failures=0, objective=-15.4]    10it [00:04,  1.89it/s, failures=0, objective=-15.4]    10it [00:04,  1.89it/s, failures=0, objective=-15.4]    11it [00:04,  2.29it/s, failures=0, objective=-15.4]    11it [00:04,  2.29it/s, failures=0, objective=-15.4]    12it [00:05,  2.08it/s, failures=0, objective=-15.4]    12it [00:05,  2.08it/s, failures=0, objective=-12.6]    13it [00:06,  1.73it/s, failures=0, objective=-12.6]    13it [00:06,  1.73it/s, failures=0, objective=-12.6]    14it [00:06,  1.90it/s, failures=0, objective=-12.6]    14it [00:06,  1.90it/s, failures=0, objective=-12.6]    15it [00:07,  1.78it/s, failures=0, objective=-12.6]    15it [00:07,  1.78it/s, failures=0, objective=-12.6]    16it [00:08,  1.33it/s, failures=0, objective=-12.6]    16it [00:08,  1.33it/s, failures=0, objective=-12.6]    17it [00:08,  1.47it/s, failures=0, objective=-12.6]    17it [00:08,  1.47it/s, failures=0, objective=-6.37]    18it [00:09,  1.92it/s, failures=0, objective=-6.37]    18it [00:09,  1.92it/s, failures=0, objective=-6.37]    19it [00:09,  1.84it/s, failures=0, objective=-6.37]    19it [00:09,  1.84it/s, failures=0, objective=-6.37]    20it [00:10,  1.56it/s, failures=0, objective=-6.37]    20it [00:10,  1.56it/s, failures=0, objective=-6.37]    21it [00:10,  2.02it/s, failures=0, objective=-6.37]    21it [00:10,  2.02it/s, failures=0, objective=-6.37]    22it [00:11,  2.16it/s, failures=0, objective=-6.37]    22it [00:11,  2.16it/s, failures=0, objective=-6.37]    23it [00:11,
[…]
