+++ docs/examples/examples_parallelism/sg_execution_times.rst
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**00:26.823** total execution time for 3 files **from examples/examples_parallelism**:
+**03:08.358** total execution time for 3 files **from examples/examples_parallelism**:
 
 .. container::
 
@@ -32,10 +32,10 @@ Computation times
    * - Example
      - Time
      - Mem (MB)
-   * - :ref:`sphx_glr_examples_examples_parallelism_plot_profile_worker_utilization.py` (``plot_profile_worker_utilization.py``)
-     - 00:26.823
-     - 0.0
    * - :ref:`sphx_glr_examples_examples_parallelism_plot_from_serial_to_parallel_hpo.py` (``plot_from_serial_to_parallel_hpo.py``)
+     - 03:08.358
+     - 0.0
+   * - :ref:`sphx_glr_examples_examples_parallelism_plot_profile_worker_utilization.py` (``plot_profile_worker_utilization.py``)
      - 00:00.000
      - 0.0
    * - :ref:`sphx_glr_examples_examples_parallelism_plot_scaling_bo.py` (``plot_scaling_bo.py``)
