### 2025-03-28 3ae0313
updated gallery example
