### 2025-03-28 c4d2a24
Refactoring parameters of CBO and deephyper.skopt.optimizer (#299)
* refactoring acq_optimizer ga with pymoo

* refactored mixedga for acq_opt

* updated skopt optimizer with ruff formatting

* removed "update_prior" from cbo and skopt space

* WIP

* updated stopping criteria for acq_opt with pymoo

* adding pydantic checks for cbo parameters, adding scheduler for surrogate model parameters in cbo

* fix tests running slower

* changing dict of pydantic object to model_dump

* updated MPIDistributedDBO to match CBO

* format _redis_storage.py

* adding delay in surrogate model params patience
[…]
