+++ src/deephyper/evaluator/callback.py
@@ -187,6 +187,9 @@ class TqdmCallback(Callback):
                     self._n_failures += 1
                 self._tqdm.set_postfix(objective=self._best_objective, failures=self._n_failures)
 
+        if self._max_evals == self._n_done:
+            self._tqdm.close()
+
 
 class SearchEarlyStopping(Callback):
     """Stop the search gracefully when it does not improve for a given number of evaluations.
