+++ src/deephyper/skopt/optimizer/acq_optimizer/pymoo_mixedga.py
@@ -1,21 +1,35 @@
+"""Mixed-Integer genetic-algorithm optimization for the acquisition function."""
+
 from collections import OrderedDict
 
 import numpy as np
-
-from ConfigSpace.forbidden import ForbiddenClause, ForbiddenRelation, ForbiddenConjunction
+from ConfigSpace.forbidden import ForbiddenClause, ForbiddenConjunction, ForbiddenRelation
 from ConfigSpace.util import deactivate_inactive_hyperparameters
-
+from pymoo.config import Config
+from pymoo.core.mixed import (
+    MixedVariableDuplicateElimination,
+    MixedVariableGA,
+    MixedVariableMating,
+)
+from pymoo.core.population import Population
 from pymoo.core.problem import ElementwiseProblem, Problem
 from pymoo.core.repair import Repair
 from pymoo.core.termination import Termination
 from pymoo.core.variable import Choice, Integer, Real
+from pymoo.optimize import minimize
 from pymoo.termination.ftol import SingleObjectiveSpaceTermination
 from pymoo.termination.max_eval import MaximumFunctionCallTermination
 from pymoo.termination.max_gen import MaximumGenerationTermination
 from pymoo.termination.robust import RobustTermination
-from pymoo.config import Config
+from sklearn.utils import check_random_state
 
 import deephyper.skopt.space as skopt_space
+
+Config.warnings["not_compiled"] = False
+
+# https://pymoo.org/customization/mixed.html
+# https://pymoo.org/interface/problem.html
+
 
 Config.warnings["not_compiled"] = False
 
@@ -49,6 +63,8 @@ def convert_space_to_pymoo_mixed(space):
 
 
 class PyMOOMixedVectorizedProblem(Problem):
+    """Pymoo mixed-integer problem definition (vectorized)."""
+
     def __init__(self, space, acq_func=None, **kwargs):
         super().__init__(vars=convert_space_to_pymoo_mixed(space), n_obj=1, **kwargs)
         self.space = space
@@ -71,6 +87,8 @@ class PyMOOMixedVectorizedProblem(Problem):
 
 
 class PyMOOMixedElementWiseProblem(ElementwiseProblem):
+    """Pymoo mixed-integer problem definition (element-wise)."""
+
     def __init__(self, space, acq_func=None, **kwargs):
         vars = convert_space_to_pymoo_mixed(space)
         super().__init__(vars=vars, n_obj=1, **kwargs)
@@ -85,18 +103,9 @@ class PyMOOMixedElementWiseProblem(ElementwiseProblem):
         out["F"] = y
 
 
-class PyMOORealProblem(Problem):
-    def __init__(self, n_var, xl, xu, acq_func, **kwargs):
-        super().__init__(n_var=n_var, n_obj=1, xl=xl, xu=xu)
-        self.acq_func = acq_func
-
-    def _evaluate(self, x, out, *args, **kwargs):
-        y = self.acq_func(x).reshape(-1)
-
-        out["F"] = y
-
-
 class DefaultMixedTermination(Termination):
+    """Pymoo custom termination criteria for mixed-integer problem."""
+
     def __init__(self, f, n_max_gen=1000, n_max_evals=100000) -> None:
         super().__init__()
         self.f = f
@@ -112,38 +121,39 @@ class DefaultMixedTermination(Termination):
 
 
 class DefaultSingleObjectiveMixedTermination(DefaultMixedTermination):
+    """Pymoo custom default single objectived mixed-integer termination criteria."""
+
     def __init__(self, ftol=1e-6, period=30, n_max_gen=1000, **kwargs) -> None:
-        f = RobustTermination(
[…]
