### 2025-03-31 #7901 @Deathn0t
Hello @jbytecode, As the review has not started yet, I updated the software version to 0.10.0 last week. Is it possible to update the review to this version? It includes a few fixes and a better choice of default parameters.
