### 2025-04-01 #7901 @toshihikoyanase
@jbytecode Thank you very much for reaching out!

I discussed with the team, including @HideakiImamura, developing the BBO library [Optuna](https://optuna.org/), and we found that one of our core committers, @y0z, is interested in participating in the review.

@y0z proposed [MoTPE](https://dl.acm.org/doi/abs/10.1145/3377930.3389817) and has also led the launch of [OptunaHub](https://hub.optuna.org/). With a deep understanding of both optimization algorithms and software development, we believe he would be well-suited for reviewing this submission.

Would it be possible to assign him as a reviewer for this JOSS submission?
