+++ src/deephyper/evaluator/storage/_mpi_win_storage.py
@@ -89,7 +89,9 @@ class MPIWinStorage(Storage):
         with self._mapping():
             self._mapping["data"][search_id]["data"][partial_id][key] = value
 
-    def store_job_in(self, job_id: Hashable, args: Tuple = None, kwargs: Dict = None) -> None:
+    def store_job_in(
+        self, job_id: Hashable, args: Tuple = None, kwargs: Dict = None
+    ) -> None:
         """Stores the input arguments of the executed job.
 
         Args:
@@ -118,7 +120,9 @@ class MPIWinStorage(Storage):
         """
         search_id, partial_id = job_id.split(".")
         with self._mapping():
-            self._mapping["data"][search_id]["data"][partial_id]["metadata"][key] = value
+            self._mapping["data"][search_id]["data"][partial_id]["metadata"][
+                key
+            ] = value
 
     def load_all_search_ids(self) -> List[Hashable]:
         """Loads the identifiers of all recorded searches.
@@ -166,7 +170,9 @@ class MPIWinStorage(Storage):
         data = self._mapping["data"][search_id]["data"][partial_id]
         return copy.deepcopy(data)
 
-    def store_search_value(self, search_id: Hashable, key: Hashable, value: Any) -> None:
+    def store_search_value(
+        self, search_id: Hashable, key: Hashable, value: Any
+    ) -> None:
         """Stores the value corresponding to key for search_id.
 
         Args:
@@ -186,7 +192,9 @@ class MPIWinStorage(Storage):
         """
         return self._mapping["data"][search_id][key]
 
-    def load_metadata_from_all_jobs(self, search_id: Hashable, key: Hashable) -> List[Any]:
+    def load_metadata_from_all_jobs(
+        self, search_id: Hashable, key: Hashable
+    ) -> List[Any]:
         """Loads a given metadata value from all jobs.
 
         Args:
