### 2025-04-07 40d66d1
add code dropdowns, small fixes
