+++ src/deephyper/evaluator/_mpi_comm.py
@@ -89,6 +89,7 @@ class MPICommEvaluator(Evaluator):
                 if self.comm.Get_rank() == self.root:
                     search_id = storage.create_new_search()
         self.comm.Barrier()
+        search_id = self.comm.bcast(search_id, self.root)
 
         super().__init__(
             run_function=run_function,
