### 2025-04-07 deephyper/deephyper#305 @y0z
JOSS Review #1
This is the first-round review for JOSS.

According to [the review checklist](https://github.com/openjournals/joss-reviews/issues/7975#issuecomment-2774649570), I carefully read the manuscript and verified the DeepHyper package.
Only minor issues were identified regarding the "general checks," "functionality," and "documentation" aspects, and I made PRs about them (https://github.com/deephyper/deephyper/issues/303, https://github.com/deephyper/deephyper/issues/304).
Regarding the "software paper" aspect, I consider the following critical issues that require a significant manuscript revision.

### Clarification of Differences from the Previous Version

I find that [the previous version of the DeepHyper paper](https://ieeexplore.ieee.org/document/8638041) was published in 2018 and provides more information (2 columns and 10 pages, including numerical results) than the current manuscript. First, the manuscript must mention and cite the previous version to avoid self-plagiarism. The differences (i.e., additional contributions) between the new manuscript and the prior version should be clarified since [JOSS does not permit repeated publication of the same work](https://joss.theoj.org/about#ethics). One idea for improving the manuscript is to enrich the contents briefly described in lines 41-57, new features published in 2022-2024, i.e., apparent updates from 2018.

### Comparison & Relation to Other Work

The manuscript contains no comparisons or discussions against related work and other commonly used packages. The authors should discuss the pros and cons of DeepHyper compared to related work (e.g., the ones mentioned in https://github.com/openjournals/joss-reviews/issues/7901#issuecomment-2729511832), and, where possible, provide some numerical results.

### Quality of Writing

- All figures must be referenced in the text. Please refer to Figs. 1, 2, and 3 appropriately. Please also refer to the figure in the mathematics section and provide a figure number and caption to it.

### Minor Issues
- It is recommended that the section "Mathematics" be renamed more informatively, as it describes DeepHyper's optimization algorithm.
- l36. assurrogate -> as surrogate
