### 2025-04-07 deephyper/deephyper#305 @Deathn0t
Hello @y0z ,

Thank you for taking the time to review our work. I'll implement the minor issues (quality of writing, typos, and #303, #304).

The [DeepHyper - 2018 paper](https://ieeexplore.ieee.org/document/8638041) you mentioned is indeed a legacy paper from this same project. This 2018 paper was not a software paper submission but an experimental scalability analysis of an early-stage version of the current `deephyper.hpo.CBO` (submission made at the time to an HPC-related conference). Since the publication of this paper, only the generic high-level logic of using Bayesian optimization remains. The algorithms and software architecture have changed (the software was not open source at the time of this publication, you can still find our oldest tag https://github.com/deephyper/deephyper/tree/0.1.1 to check the difference with the current version) . We can add the citation in the JOSS paper, but we did not want to avoid confusing the user with the legacy results found in this paper (when the recent version is significantly better). To resolve this, we suggest adding the reference to the 2018 paper to our JOSS submission with a synthethic explanation of the legacy from this initial work.

The papers referenced in our submission all include comparison to other works. However, we do not have such plot in the content of the JOSS submission. While looking at other JOSS submissions [example of accepted paper](https://github.com/openjournals/joss-papers/blob/5d949bb3f86acad85820980bb9f9825af286d453/joss.07861/10.21105.joss.07861.pdf) it seemed that the content of the paper was synthetic, so we did the same. If you think we should add a comparison plot in our submission, let us know we can work on it. A comparison could include "Software Feature Matrix", "Performance Benchmark for Black-Box Benchmarks".

Best regards,
Romain
