### 2025-04-07 deephyper/deephyper#305 @y0z
Thank you for your response.

> To resolve this, we suggest adding the reference to the 2018 paper to our JOSS submission with a synthetic explanation of the legacy from this initial work.

OK. It will resolve the issue.

> The papers referenced in our submission all include comparison to other works... A comparison could include "Software Feature Matrix", "Performance Benchmark for Black-Box Benchmarks".

Numerical results can be optional, but the manuscript should cover the following points in the review criteria.

**A statement of need:** Does the paper have a section titled 'Statement of need' that *clearly states* what problems the software is designed to solve, who the target audience is, and *its relation to other work*?

**State of the field:** Do the authors *describe how this software compares to other commonly-used packages*?

Regards,

Yoshihiko
