### 2025-04-08 7be7649
adding optimization for RMA window allocation
