+++ src/deephyper/evaluator/storage/_mpi_win_mutable_mapping.py
@@ -49,15 +49,20 @@ class MPIWinMutableMapping(MutableMapping):
 
         # Allocate memory (works on multiple nodes)
         logging.info("Allocating MPI.Win ...")
-        self.win = MPI.Win.Allocate(size, 1, comm=comm)
-        # TODO: shake if comm is on 1 node or multiple nodes to use:
-        # TODO: Allocate_shared + Shared_query instead
-        # self.win = MPI.Win.Allocate_shared(size, 1, comm=comm)
-        # buf, itemsize = self.win.Shared_query(self.root)
-        # self.shared_memory = np.ndarray(buffer=buf, dtype=np.byte, shape=(size,))
-        logging.info("MPI.Win allocated")
 
-        self.shared_memory = np.empty((size,), dtype=np.byte)
+        # If all processes are in the local shared comm then we
+        # can use shared memory
+        local_comm = self.comm.Split_type(MPI.COMM_TYPE_SHARED)
+        if local_comm.Get_size() == self.comm.Get_size():
+            logging.info("Using MPI.Win.Allocate_shared")
+            self.win = MPI.Win.Allocate_shared(size, 1, comm=comm)
+            buf, itemsize = self.win.Shared_query(self.root)
+            self.shared_memory = np.ndarray(buffer=buf, dtype=np.byte, shape=(size,))
+        else:
+            logging.info("Using MPI.Win.Allocate")
+            self.win = MPI.Win.Allocate(size, 1, comm=comm)
+            self.shared_memory = np.empty((size,), dtype=np.byte)
+        logging.info("MPI.Win allocated")
 
         if default_value is None:
             self.local_dict = {}
@@ -86,9 +91,13 @@ class MPIWinMutableMapping(MutableMapping):
         try:
             self.win.Get(self.shared_memory, target_rank=self.root)
             self.win.Flush(self.root)
-            size = int.from_bytes(self.shared_memory[: self.HEADER_SIZE], byteorder="big")
+            size = int.from_bytes(
+                self.shared_memory[: self.HEADER_SIZE], byteorder="big"
+            )
             if size > 0:
-                raw_data = self.shared_memory[self.HEADER_SIZE : self.HEADER_SIZE + size].tobytes()
+                raw_data = self.shared_memory[
+                    self.HEADER_SIZE : self.HEADER_SIZE + size
+                ].tobytes()
                 self.local_dict = pickle.loads(raw_data)
             else:
                 self.local_dict = {}
@@ -182,7 +191,9 @@ class MPIWinMutableMapping(MutableMapping):
 
     def session_start(self, read_only: bool = False):
         if self._session_is_started:
-            raise RuntimeError("A session has already been started without being finished!")
+            raise RuntimeError(
+                "A session has already been started without being finished!"
+            )
 
         self._session_is_started = True
         self._session_is_read_only = read_only
