### 2025-04-08 c383b37
test update format with cited figures
