+++ paper.md
@@ -39,8 +39,9 @@ Machine learning models are increasingly applied across scientific disciplines, 
 
 # Statement of need
 
-`DeepHyper` is a Python package for parallel hyperparameter optimization or neural architecture search. It provides
-access to a variety of asynchronous parallel black-box optimization algorithms `deephyper.hpo`. The 
+`DeepHyper` is a Python package for parallel hyperparameter optimization or neural architecture search.
+The project started in 2018 [@balaprakash2018deephyper] with a focus on making Bayesian optimization more efficient on high-performance computing clusters.
+It provides access to a variety of asynchronous parallel black-box optimization algorithms `deephyper.hpo`. The 
 software offers a variety of parallel programming backends such as Asyncio, threading, processes, Ray, and MPI `deehyper.evaluator`. The hyperparameter optimization can be single or multi-objective, composed of mixed variables,
 using explicit or hidden constraints, and benefit from early-discarding strategies `deephyper.stopper`. Leveraging the results of hyperparameter optimization or neural architecture search it provides parallel ensemble algorithms `deephyper.ensemble` that can help improve accuracy or quantify disentangled predictive uncertainty.
 
@@ -49,18 +50,30 @@ using explicit or hidden constraints, and benefit from early-discarding strategi
 `DeepHyper` was designed to help research in the field of automated machine learning and also to be used out-of-the box
 in scientific projects where learning workflows are being developed.
 
-# Mathematics
+# Related Works
 
-`DeepHyper` main hyperparameter optimization algorithm is based on Bayesian optimization.
+## Feature Matrix
 
-It can work with mixed-integer problems (i.e., including real, discrete, categorical values), with explicit or implicit constraints (e.g., $x_0 < x_1$, unexpected out-of-memory error), and stochastic objective functions $Y = f(x)$ where $Y$ is a random variable, $x$ is a vector of input hyperparameters and $f$ is the objective function.
+...
 
-The Bayesian optimization of the `DeepHyper` relies by default on Extremely Randomized Forest assurrogate model to estimate $E_Y[Y|X=x]$.
-Extremely Randomized Forest are a kind of Random Forest where the split decision involves a random process for each newly created node of a tree.
-It provides smoother epistemic uncertainty estimates compared to usual Random Forests that use a deterministic "best" split decision.
+## Black-box Optimization Benchmarks
 
-![Uncertainty for Random Forest (Best Split)](figures/random_forest_best_split.png){width=49%}
-![Uncertainty for Extremely Randomized Forest (Random Split)](figures/random_forest_random_split.png){width=49%}
+...
+
+
+# Black-box Optimization Methods
+
+The algorithms used for hyperparameter optimization are black-box optimization algorithms for mixed search spaces.
+A mixed search space, can be composed of real, discrete, or categorical (nomimal or ordinal) values.
+The search space can also include constraints (e.g., explicit such as $x_0 < x_1$, or implicit such as "unexpected out-of-memory error").
+The objective function can be stochastic $Y = f(x)$ where $Y$ is a random variable, $x$ is a vector of input hyperparameters and $f$ is the objective function. `DeepHyper` main hyperparameter optimization algorithm is based on Bayesian optimization.
+
+The Bayesian optimization of `DeepHyper` relies by default on Extremely Randomized Forest as surrogate model to estimate $E_Y[Y|X=x]$.
+Extremely Randomized Forest [@geurts2006extremely] are a kind of Random Forest algorithm where the split decision involves a random process for each newly created node of a tree.
[…]
