+++ paper.bib
@@ -132,3 +132,12 @@
   doi       = {10.1016/j.neucom.2024.127964},
   publisher = {Elsevier}
 }
+
+@article{geurts2006extremely,
+  title     = {Extremely randomized trees},
+  author    ={Geurts, Pierre and Ernst, Damien and Zehenkel, Louis},
+  journal   = {Machine Learning},
+  year      = {2006},
+  doi       = {10.1007/s10994-006-6226-1},
+  publisher = {Springer}
+}
