### 2025-04-08 6ce7077
Optimization for RMA window memory allocation (#306)
