+++ paper.md
@@ -69,15 +69,15 @@ The search space can also include constraints (e.g., explicit such as $x_0 < x_1
 The objective function can be stochastic $Y = f(x)$ where $Y$ is a random variable, $x$ is a vector of input hyperparameters and $f$ is the objective function. `DeepHyper` main hyperparameter optimization algorithm is based on Bayesian optimization.
 
 The Bayesian optimization of `DeepHyper` relies by default on Extremely Randomized Forest as surrogate model to estimate $E_Y[Y|X=x]$.
-Extremely Randomized Forest [@geurts2006extremely] are a kind of Random Forest algorithm where the split decision involves a random process for each newly created node of a tree.
-It provides smoother epistemic uncertainty estimates with an increasing number of trees (Figure \ref{random-forest-random-split}) compared to usual Random Forests that use a deterministic "best" split decision (Figure \ref{random-forest-best-split}).
+Extremely Randomized Forest [@geurts2006extremely] are a kind of Randomized Forest algorithm where the split decision involves a random process for each newly created node of a tree.
+It provides smoother epistemic uncertainty estimates with an increasing number of trees (Figure \ref{randomsplit}) compared to usual Random Forests that use a deterministic "best" split decision (Figure \ref{bestsplit}).
 
-![Uncertainty for Random Forest (Best Split)](figures/random_forest_best_split.png){width=45%, label="random-forest-best-split"}
+![Uncertainty for Random Forest (Best Split)](figures/random_forest_best_split.png){width=45%,label="bestsplit"}
 
-![Uncertainty for Extremely Randomized Forest (Random Split)](figures/random_forest_random_split.png){width=45%, label="random-forest-random-split"}
+![Uncertainty for Extremely Randomized Forest (Random Split)](figures/random_forest_random_split.png){width=45%,label="randomsplit"}
 
-Then, a custom acquisition function `UCBd`, focuses on the epistemic uncertainty of this surrogate for improved efficiency.
-It is also combined with a periodic exponential decay (impacting exploration-exploitation parameters of BO) to escape local solutions [@egele2023asynchronous].
+Then, a custom acquisition function `UCBd`, focuses on the epistemic uncertainty of this surrogate (purple area in Figure 4) for improved efficiency.
+It is also combined with a periodic exponential decay (Figure 5, impacting exploration-exploitation parameters of BO) to escape local solutions [@egele2023asynchronous].
 
 ![Periodic Exponential Decay for Bayesian Optimization](figures/example-exp-decay.jpg){width=49%}
 
