+++ paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 
 ---
 
-![DeepHyper Logo](figures/logo-deephyper-nobg.png){width=40%}
+![](figures/logo-deephyper-nobg.png){width=40%}
 
 # Summary
 
@@ -43,22 +43,153 @@ Machine learning models are increasingly applied across scientific disciplines, 
 The project started in 2018 [@balaprakash2018deephyper] with a focus on making Bayesian optimization more efficient on high-performance computing clusters.
 It provides access to a variety of asynchronous parallel black-box optimization algorithms `deephyper.hpo`. The 
 software offers a variety of parallel programming backends such as Asyncio, threading, processes, Ray, and MPI `deehyper.evaluator`. The hyperparameter optimization can be single or multi-objective, composed of mixed variables,
-using explicit or hidden constraints, and benefit from early-discarding strategies `deephyper.stopper`. Leveraging the results of hyperparameter optimization or neural architecture search it provides parallel ensemble algorithms `deephyper.ensemble` that can help improve accuracy or quantify disentangled predictive uncertainty.
+using explicit or hidden constraints, and benefit from early-discarding strategies `deephyper.stopper`. Leveraging the results of hyperparameter optimization or neural architecture search it provides parallel ensemble algorithms `deephyper.ensemble` that can help improve accuracy or quantify disentangled predictive uncertainty. A diagram of our software architecture is shown in Figure 1.
 
-![DeepHyper Software Architecture](figures/deephyper-architecture.png){width=60%}
+![Software Architecture](figures/deephyper-architecture.png){width=60%}
 
 `DeepHyper` was designed to help research in the field of automated machine learning and also to be used out-of-the box
 in scientific projects where learning workflows are being developed.
 
 # Related Works
 
-## Feature Matrix
+Numerous software packages now exist for hyperparameter optimization (HPO), including:
 
-...
+- [HyperMapper]()
+- [Optuna](https://optuna.readthedocs.io)
+- [Hyperopt]()
+- [BoTorch]()
+- [OpenBox]()
+- [HpBandSter]()
+- [SMAC](https://automl.github.io/SMAC3)
+
+These tools differ in scope and design: some are research-oriented (e.g., SMAC), while others prioritize production-readiness and usability (e.g., Optuna). In this section, we focus our comparison on SMAC and Optuna, which are representative of these two directions.
+
+**DeepHyper** is designed to maximize HPO efficiency across a wide range of parallelization scales, from sequential (single-core) runs to massively parallel evaluations on high-performance computing (HPC) systems with thousands of cores.
+
+## Feature Comparison
+
+While the feature matrix below provides a high-level overview, it necessarily simplifies some nuanced implementation differences.
+
+### Hyperparameter Optimization Capabilities
+
+
+| Feature | DeepHyper       | Optuna | SMAC3 |
+|---|-----------------|--------|-------|
+| Single-objective       | 🟩 | 🟩 | 🟩    |
+| Multi-objective        | 🟩 | 🟩 | 🟩    |
+| Early discarding       | 🟩 | 🟩 | 🟩    |
+| Fault tolerance        | 🟩 | 🟩 | 🟥    |
+| Transfer learning      | 🟩 | 🟩 | 🟧    |
+| Ensemble construction  | 🟩 | 🟥 | 🟥    |
+| Visualization          | 🟧 | 🟩 | 🟥    |
+
+Absolutely! Here's an improved version of your text that adopts a **factual, scientific, formal, and synthetic** tone while preserving the important content. I've also polished the grammar and structure for clarity and readability.
+
+---
+
+# Related Work
+
[…]
