### 2025-04-08 deephyper/deephyper#305 @Deathn0t
Hello @Deathn0t , I am attaching an updated draft that follows your suggestions. Let me know if this goes in the direction you expect.

[paper.pdf](https://github.com/user-attachments/files/19647398/paper.pdf)
