### 2025-04-08 ea517be
matplotlic rc do not use LaTeX by default
