### 2025-04-09 947ae26
update documentation of profile decorator
