### 2025-04-14 69cd83a
fix colab dependency errors
