### 2025-04-14 ad70ced
one line pip install
