### 2025-04-15 a37fb3e
use latex checkmarks in first table
