### 2025-04-15 83918ca
use latex checkmarks in second table
