### 2025-04-15 deephyper/deephyper#305 @y0z
Thank you for the update. Although LLM can help improve the quality of your manuscript, please double-check it before resubmitting to detect apparent flaws.

I confirmed that the related work section was added to cover comparisons with other work.
Then, I leave some additional comments.

1. Devoting much space to computational experiments is not in line with JOSS's direction. So, the section on black-box optimization benchmarks can be omitted as it does not explain the benchmarking problem and discuss the results obtained, i.e., not helpful for readers.

2. Please provide each table/figure's number, title, and caption and refer to each table in the text, such as "... is shown in Table/Figure X." For example, logo-deephyper-nobg.png does not have its caption nor be referred to in the text. A feature comparison table does not have the number, title, and caption and is not referred to in the text.

You can find how to add a caption to a table from a paper published in JOSS.
*Example*:
- Paper: https://joss.theoj.org/papers/10.21105/joss.04684
- Markdown: https://github.com/easystats/datawizard/blob/main/paper/JOSS_files/paper.md?plain=1#L71-L84

3. Please cite the paper for each OSS in the related work section (i.e., adding them to references).
- HyperMapper: https://ieeexplore.ieee.org/document/9124618
- Optuna: https://dl.acm.org/doi/10.1145/3292500.3330701
- Hyperopt: https://iopscience.iop.org/article/10.1088/1749-4699/8/1/014008
- BoTorch: https://proceedings.neurips.cc/paper/2020/hash/f5b1b89d98b7286673128a5fb112cb9a-Abstract.html
- OpenBox: http://jmlr.org/papers/v25/23-0537.html
- HpBandSter: https://proceedings.mlr.press/v80/falkner18a.html
- SMAC: https://jmlr.org/papers/v23/21-0888.html
