### 2025-04-15 deephyper/deephyper#305 @Deathn0t
Hello @y0z ,

1. I will add details about the quantitative benchmarks as the results are available. It is maybe better to improve the presentation rather than remove it, maybe I could keep fewer figures (aggregating them would hide the specificities of each benchmark).

2. I will implement these points. For the figure labels, the logo omits the title on purpose, as it is used as a logo. For the internal references, I tried previously to use [this JOSS feature](https://joss.readthedocs.io/en/latest/paper.html#tables-and-figures), but the compilation did not work. I will try a different method I just found [in this example](https://joss.readthedocs.io/en/latest/example_paper.html). Thanks for the reference to table captions.

Thanks for listing the DOIs. I was planning to add links to each software's documentation but I will replace by official references.
