### 2025-04-16 efa7e84
Improve hpo text classification tutorial (#302)
* add code dropdowns, small fixes

* fix colab dependency errors

* one line pip install

* remove unused import

---------

Co-authored-by: Brett Eiffert <eiffertbc@ornl.gov>
Co-authored-by: Romain Egele <romainegele@gmail.com>
