+++ docs/examples/examples_hpo/plot_hpo_text_classification.rst
@@ -38,59 +38,72 @@ This tutorial is based on materials from the Pytorch Documentation: `Text classi
     pip install ray
     pip install torch torchtext torchdata
 
-.. GENERATED FROM PYTHON SOURCE LINES 24-26
+.. GENERATED FROM PYTHON SOURCE LINES 24-28
 
 Imports
 -------
 
-.. GENERATED FROM PYTHON SOURCE LINES 28-44
+All imports used in the tutorial are declared at the top of the file.
 
-.. code-block:: Python
+.. GENERATED FROM PYTHON SOURCE LINES 28-47
 
-    import ray
-    import json
-    import pandas as pd
-    from functools import partial
+.. dropdown:: Code (Imports)
 
-    import torch
-
-    from torchtext.data.utils import get_tokenizer
-    from torchtext.data.functional import to_map_style_dataset
-    from torchtext.vocab import build_vocab_from_iterator
-
-    from torch.utils.data import DataLoader
-    from torch.utils.data.dataset import random_split
-
-    from torch import nn
+    .. code-block:: Python
 
 
+        import ray
+        import json
+        import pandas as pd
+        from functools import partial
+
+        import torch
+
+        from torchtext.data.utils import get_tokenizer
+        from torchtext.data.functional import to_map_style_dataset
+        from torchtext.vocab import build_vocab_from_iterator
+        from torchtext.datasets import AG_NEWS
+
+        from torch.utils.data import DataLoader
+        from torch.utils.data.dataset import random_split
+
+        from torch import nn
 
 
 
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 45-48
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 48-51
 
 .. note::
-  The following can be used to detect if <b>CUDA</b> devices are available on the current host. Therefore, this notebook will automatically adapt the parallel execution based on the ressources available locally. However, it will not be the case if many compute nodes are requested.
+  The following can be used to detect if **CUDA** devices are available on the current host. Therefore, this notebook will automatically adapt the parallel execution based on the ressources available locally. However, it will not be the case if many compute nodes are requested.
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 50-53
+.. GENERATED FROM PYTHON SOURCE LINES 53-54
 
-.. code-block:: Python
+If GPU is available, this code will enabled the tutorial to use the GPU for pytorch operations.
 
-    is_gpu_available = torch.cuda.is_available()
-    n_gpus = torch.cuda.device_count()
+.. GENERATED FROM PYTHON SOURCE LINES 55-60
 
+.. dropdown:: Code (Code to check if using CPU or GPU)
 
-
-
-
+    .. code-block:: Python
 
 
+        is_gpu_available = torch.cuda.is_available()
+        n_gpus = torch.cuda.device_count()
 
-.. GENERATED FROM PYTHON SOURCE LINES 54-59
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 61-66
 
 The dataset
 -----------
@@ -98,36 +111,37 @@ The dataset
 The torchtext library provides a few raw dataset iterators, which yield the raw text strings. For example, the :code:`AG_NEWS` dataset iterators yield the raw data as a tuple of label and text. It has four labels (1 : World 2 : Sports 3 : Business 4 : Sci/Tec).
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 61-79
+.. GENERATED FROM PYTHON SOURCE LINES 66-84
 
-.. code-block:: Python
+.. dropdown:: Code (Loading the data)
 
-    from torchtext.datasets import AG_NEWS
+    .. code-block:: Python
 
-    def load_data(train_ratio, fast=False):
[…]
