### 2025-04-21 deephyper/deephyper#313 @jungtaekkim
JOSS Review: Manuscript
Hi the authors of deephyper,

Thank you for your submission. This is a valuable contribution, and I appreciate the effort that went into developing the software. I would like to provide the following comments on your JOSS submission:

1. The current version appears too sparse in its presentation. For example, Figures 3 to 11 could be consolidated into a single figure or split into just one or two comprehensive figures. Similarly, Figures 12 and 13 can be merged into one figure for better readability and compactness.
2. The manuscript is currently lacking many relevant citations. While JOSS places more emphasis on the software contribution, the academic context still matters. All related software mentioned in the Related Work section should be properly cited. Please also ensure that all relevant prior work is included and appropriately referenced.
3. A more thorough discussion on the software development aspects would strengthen the manuscript. This includes details on software dependencies, maintenance strategy, and example usage or sample code to help readers understand and adopt the software more easily.
4. The Ax platform (https://github.com/facebook/Ax) should also be discussed in the Related Work section, as it is highly relevant to the domain.
5. Please provide captions for all tables. Additionally, the comparison tables would benefit from including a broader range of related software for a more comprehensive evaluation.
