### 2025-04-21 deephyper/deephyper#314 @jungtaekkim
JOSS Review: Codebase
Hi the authors of deephyper,

Thank you for your submission. Following the issue on your manuscript (https://github.com/deephyper/deephyper/issues/313), I would like to provide the following comments on the codebase of deephyper:

1. The codebase appears to be well-maintained and thoughtfully developed. It is great work.
2. I would suggest considering whether the `src/deephyper` directory structure is necessary. It may be cleaner and more conventional to place the code directly under `deephyper`, without the `src` directory.
3. In the following example https://github.com/deephyper/deephyper/blob/1d4f590f6f627dec6155b4d9d44c79c9c039cfbb/examples/examples_hpo/plot_hpo_text_classification.py#L150, the line can be simplified to `super().__init__()` for clarity and conciseness.
