### 2025-04-21 deephyper/deephyper#314 @Deathn0t
Hello @jungtaekkim , thanks for your review.

About the `src/deephyper`, we would prefer to keep the "src-layout" (it has now become one of the standard advised practice) as it avoids accidental imports such as importing the "local possibly development" version of the package (as described at [src layout vs flat layout](https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/)).
