+++ paper.bib
@@ -125,7 +125,7 @@
 
 @article{egele2024unreasonable,
   title     = {The unreasonable effectiveness of early discarding after one epoch in neural network hyperparameter optimization},
-  author    ={Egele, Romain and Mohr, Felix and Viering, Tom and Balaprakash, Prasanna},
+  author    = {Egele, Romain and Mohr, Felix and Viering, Tom and Balaprakash, Prasanna},
   journal   = {Neurocomputing},
   pages     = {127964},
   year      = {2024},
@@ -135,9 +135,101 @@
 
 @article{geurts2006extremely,
   title     = {Extremely randomized trees},
-  author    ={Geurts, Pierre and Ernst, Damien and Zehenkel, Louis},
+  author    = {Geurts, Pierre and Ernst, Damien and Zehenkel, Louis},
   journal   = {Machine Learning},
   year      = {2006},
   doi       = {10.1007/s10994-006-6226-1},
   publisher = {Springer}
+}
+
+@inproceedings{hypermapper2019,
+  author    = {Nardi, Luigi and Souza, Artur and Koeplinger, David and Olukotun, Kunle},
+  booktitle = {2019 IEEE 27th International Symposium on Modeling, Analysis, and Simulation of Computer and Telecommunication Systems (MASCOTS)},
+  title     = {HyperMapper: a Practical Design Space Exploration Framework},
+  year      = {2019},
+  volume    = {},
+  number    = {},
+  pages     = {425-426},
+  keywords  = {Optimization;Software;Space exploration;Field programmable gate arrays;Computational modeling;Tuning;Response surface methodology},
+  doi       = {10.1109/MASCOTS.2019.00053}
+}
+
+@inproceedings{optuna2019,
+  author    = {Akiba, Takuya and Sano, Shotaro and Yanase, Toshihiko and Ohta, Takeru and Koyama, Masanori},
+  title     = {Optuna: A Next-generation Hyperparameter Optimization Framework},
+  year      = {2019},
+  isbn      = {9781450362016},
+  publisher = {Association for Computing Machinery},
+  address   = {New York, NY, USA},
+  url       = {https://doi.org/10.1145/3292500.3330701},
+  doi       = {10.1145/3292500.3330701},
+  booktitle = {Proceedings of the 25th ACM SIGKDD International Conference on Knowledge Discovery \& Data Mining},
+  pages     = {2623–2631},
+  numpages  = {9},
+  keywords  = {machine learning system, hyperparameter optimization, black-box optimization, Bayesian optimization},
+  location  = {Anchorage, AK, USA},
+  series    = {KDD '19}
+}
+
+@inproceedings{hyperopt2013,
+  author    = {Bergstra, J. and Yamins, D. and Cox, D. D.},
+  title     = {Making a science of model search: hyperparameter optimization in hundreds of dimensions for vision architectures},
+  year      = {2013},
+  publisher = {JMLR.org},
+  booktitle = {Proceedings of the 30th International Conference on International Conference on Machine Learning - Volume 28},
+  pages     = {I–115–I–123},
+  location  = {Atlanta, GA, USA},
+  series    = {ICML'13},
+  doi       = {10.5555/3042817.3042832}
+}
+
+@article{hyperopt2015,
+  title     = {Hyperopt: a python library for model selection and hyperparameter optimization},
+  author    = {Bergstra, James and Komer, Brent and Eliasmith, Chris and Yamins, Dan and Cox, David D},
+  journal   = {Computational Science \& Discovery},
+  volume    = {8},
+  number    = {1},
+  pages     = {014008},
[…]
