+++ paper.md
@@ -54,13 +54,12 @@ in scientific projects where learning workflows are being developed.
 
 Numerous software packages now exist for hyperparameter optimization (HPO), including:
 
-- [HyperMapper]()
-- [Optuna](https://optuna.readthedocs.io)
-- [Hyperopt]()
-- [BoTorch]()
-- [OpenBox]()
-- [HpBandSter]()
-- [SMAC](https://automl.github.io/SMAC3)
+- BoTorch (and Ax platform) [@botorch2020], ([doc.](https://botorch.org))
+- HyperMapper [@hypermapper2019], ([doc.](https://github.com/luinardi/hypermapper))
+- Hyperopt [@hyperopt2013,@hyperopt2015], ([doc.](http://hyperopt.github.io/hyperopt/))
+- OpenBox [@openbox2024], ([doc.](https://open-box.readthedocs.io))
+- Optuna [@optuna2019], ([doc.](https://optuna.readthedocs.io))
+- SMAC3 [@smac32022], ([doc.](https://automl.github.io/SMAC3))
 
 These tools differ in scope and design: some are research-oriented (e.g., SMAC), while others prioritize production-readiness and usability (e.g., Optuna). In this section, we focus our comparison on SMAC and Optuna, which are representative of these two directions.
 
