### 2025-04-21 deephyper/deephyper#314 @wigging
Regarding the reviewer's comment in item number 2:

> I would suggest considering whether the src/deephyper directory structure is necessary. It may be cleaner and more conventional to place the code directly under deephyper, without the src directory.

There are many reasons to use `src/deephyper` instead of just `deephyper` for the directory structure. This is often referred to as the src layout vs flat layout.

Deephyper adheres to packaging practices suggested by the [Python Packaging User Guide](https://packaging.python.org/en/latest/) which is maintained by the Python Packaging Authority. First, the guide's [Packaging Python Projects](https://packaging.python.org/en/latest/tutorials/packaging-projects/) tutorial uses the src layout, not the flat layout. Second, the guide provides an entire discussion about the [src vs flat layout](https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/). I won't go into all the details but basically the src layout requires the installation of the package to run its code which the flat layout does not. The src layout makes sure that only the required package files are importable and not extraneous files like the README.md, LICENSE.md, ruff.toml, etc. which would be included with a flat layout.

Deephyper uses [pytest](https://docs.pytest.org/en/stable/index.html) for testing. In the [Good Integration Practices](https://docs.pytest.org/en/stable/explanation/goodpractices.html) section of the pytest documentation, it strongly suggests using the src layout. This ensures that the tests are run on the installed version of the package and not directly on the code files themselves. Also see this [Packaging a python library](https://blog.ionelmc.ro/2014/05/25/python-packaging/) article about the many reasons to use the src layout.

Modern Python packages should use the src layout for the reasons mentioned above. Since deephyper is a modern Python package then it should use the more conventional src layout (`src/deephyper`) and not the antiquated flat layout (`deephyper`). In addition to @Deathn0t's comment above, I disagree with the reviewer's comment to use the flat layout for the deephyper package.
