### 2025-04-21 deephyper/deephyper#313 @wigging
Related to the third item:

> 3. A more thorough discussion on the software development aspects would strengthen the manuscript. This includes details on software dependencies, maintenance strategy, and example usage or sample code to help readers understand and adopt the software more easily.

The [deephyper documentation](https://deephyper.readthedocs.io/en/latest/) has a Developer's Guide section (see bottom of left navigation column). This section contains information about software development such as linting, formatting, testing, and overall software architecture. Is this the type of information you are looking for regarding software development aspects? Are you suggesting that the developer information in the documentation needs to be summarized in the JOSS manuscript too?
