### 2025-04-21 deephyper/deephyper#313 @jungtaekkim
> Is this the type of information you are looking for regarding software development aspects?

Yes, it is what I am looking for.

> Are you suggesting that the developer information in the documentation needs to be summarized in the JOSS manuscript too?

Yes, but if it is too long, you can mention the location of this information (with a link) in the manuscript.
