+++ .github/workflows/joss.yml
@@ -1,9 +1,10 @@
 name: JOSS Draft PDF
 
-on: 
+on:
   push:
-    branches:
-      - joss
+    paths:
+      - joss/**
+      - .github/workflows/joss.yml
 
 jobs:
   paper:
@@ -19,7 +20,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: paper.md
+          paper-path: joss/paper.md
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
@@ -27,4 +28,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper.pdf
+          path: joss/paper.pdf
