### 2025-04-21 deephyper/deephyper#314 @bretteiffert
Changes made for item 3 @Deathn0t
