+++ joss/paper.md
@@ -155,6 +155,10 @@ The ensemble strategies is modular to allow: exploring models tested during hype
 
 A transfer-learning strategy for hyperparameter optimization [@dorier2022transferlearning] is also available. This strategy used to be based on variational auto-encoders for tabular data. It is now based on Gaussian mixture-model for tabular data.
 
+# Software Development
+
+The DeepHyper package adheres to best practices in the Python community by following the [PEP 8 style guide](https://pep8.org) for Python naming and formatting conventions. The layout and configuration of the package follows suggestions made by the [Python Packaging User Guide](https://packaging.python.org/en/latest/) which is maintained by the Python Packaging Authority. A `pyproject.toml` file provides all configuration settings and meta data for developing and publishing the DeepHyper package. The [ruff tool](https://astral.sh/ruff) is used to enforce style and format conventions during code development and during continuous integration checks via GitHub Actions. Unit tests are also conducted in the CI workflow with the [pytest](https://docs.pytest.org/en/stable/) framework. Conda environments and standard Python virtual environments are supported by the project to ensure consistent development environments. Documentation for the package is generated with the [Sphinx application](https://www.sphinx-doc.org/en/master/). See the Developer's Guide section in the DeepHyper [documentation](https://deephyper.readthedocs.io/en/stable/) for contributing guidelines and more software development information.
+
 # Acknowledgements
 
 We acknowledge contributions from Misha Salim, Romit Maulik, Venkat Vishwanath, Stefan Wild, Joceran Gouneau, Dipendra Jha, Kyle Felker, Matthieu Dorier, Felix Perez, Bethany Lush, Gavin M. Wiggins, Tyler H. Chang, Yixuan Sun, Shengli Jiang, rmjcs2020, Albert Lam, Taylor Childers, Z223I, Zachariah Carmichael, Hongyuan Liu, Sam Foreman, Akalanka Galappaththi, Brett Eiffert, Sun Haozhe, Sandeep Madireddy, Adrian Perez Dieguez, and Nesar Ramachandra. We also would like to thank Prof. Isabelle Guyon for her guidance on the machine learning methodologies developed in this software.
