### 2025-04-22 ee97f90
Add software development section (#317)
* Fix spelling and spacing

* Add software development section
