+++ joss/paper.md
@@ -23,12 +23,11 @@ authors:
   - name: Brett Eiffert
     affiliation: 1
 affiliations:
- - name: Oak Ridge National Laboratory, Tenesse, United States
-   index: 1
-   ror: 01qz5mb56
+  - name: Oak Ridge National Laboratory, Oak Ridge, TN, United States
+    index: 1
+    ror: 01qz5mb56
 date: 3 February 2025
 bibliography: paper.bib
-
 ---
 
 # Summary
@@ -37,16 +36,11 @@ Machine learning models are increasingly applied across scientific disciplines, 
 
 # Statement of need
 
-`DeepHyper` is a Python package for parallel hyperparameter optimization or neural architecture search.
-The project started in 2018 [@balaprakash2018deephyper] with a focus on making Bayesian optimization more efficient on high-performance computing clusters.
-It provides access to a variety of asynchronous parallel black-box optimization algorithms `deephyper.hpo`. The 
-software offers a variety of parallel programming backends such as Asyncio, threading, processes, Ray, and MPI `deehyper.evaluator`. The hyperparameter optimization can be single or multi-objective, composed of mixed variables,
-using explicit or hidden constraints, and benefit from early-discarding strategies `deephyper.stopper`. Leveraging the results of hyperparameter optimization or neural architecture search it provides parallel ensemble algorithms `deephyper.ensemble` that can help improve accuracy or quantify disentangled predictive uncertainty. A diagram of our software architecture is shown in Figure 1.
+`DeepHyper` is a Python package for parallel hyperparameter optimization or neural architecture search. The project started in 2018 [@balaprakash2018deephyper] with a focus on making Bayesian optimization more efficient on high-performance computing clusters. It provides access to a variety of asynchronous parallel black-box optimization algorithms `deephyper.hpo`. The software offers a variety of parallel programming backends such as Asyncio, threading, processes, Ray, and MPI `deehyper.evaluator`. The hyperparameter optimization can be single or multi-objective, composed of mixed variables, using explicit or hidden constraints, and benefit from early-discarding strategies `deephyper.stopper`. Leveraging the results of hyperparameter optimization or neural architecture search it provides parallel ensemble algorithms `deephyper.ensemble` that can help improve accuracy or quantify disentangled predictive uncertainty. A diagram of our software architecture is shown in Figure 1.
 
 ![Software Architecture](figures/deephyper-architecture.png){width=60%}
 
-`DeepHyper` was designed to help research in the field of automated machine learning and also to be used out-of-the box
-in scientific projects where learning workflows are being developed.
+`DeepHyper` was designed to help research in the field of automated machine learning and also to be used out-of-the box in scientific projects where learning workflows are being developed.
 
 # Related Work
 
@@ -118,8 +112,7 @@ The main difference between DeepHyper, Optuna and SMAC related to parallelizatio
 
 **Centralized optimization**: DeepHyper's allows to run centralized optimization, including $1$ master running the optimizationa and $N$ workers evaluating hyperparameter configurations.
 
-**Decentralized optimization**: DeepHyper's allows to run decentralized optimization,
-including $N$ workers, each running centralized optimization.
+**Decentralized optimization**: DeepHyper's allows to run decentralized optimization, including $N$ workers, each running centralized optimization.
 
[…]
