### 2025-04-22 deephyper/deephyper#313 @wigging
@jungtaekkim I added a Software Development section to the manuscript. This should address your comments for item 3.
