### 2025-04-22 deephyper/deephyper#313 @jungtaekkim
Thank you for your prompt response.

I think the current version looks better.

However, in Line 124, a handle in "purple area in ??" might be wrong.  Please fix it.
