### 2025-04-22 deephyper/deephyper#313 @Deathn0t
yes! I am working on that, I tried using the `\label` and `\autoref` but it does not seem to work. so I'll have to put the number of figures by hand.
