### 2025-04-23 22cd570
Joss tutorial cleanup (#316)
* JOSS edit

* JOSS edit

* add files from new run

---------

Co-authored-by: Brett Eiffert <eiffertbc@ornl.gov>
