+++ docs/examples/examples_hpo/plot_hpo_text_classification.rst
@@ -29,23 +29,21 @@ In this tutorial we present how to use hyperparameter optimization on a text cla
 **Reference**:
 This tutorial is based on materials from the Pytorch Documentation: `Text classification with the torchtext library <https://pytorch.org/tutorials/beginner/text_sentiment_ngrams_tutorial.html>`_
 
-.. GENERATED FROM PYTHON SOURCE LINES 15-21
+.. GENERATED FROM PYTHON SOURCE LINES 15-19
 
 .. code-block:: bash
 
     %%bash
-    pip install deephyper
-    pip install ray
-    pip install torch torchtext torchdata
+    pip install deephyper ray numpy==1.26.4 torch torchtext==0.17.2 torchdata==0.7.1 'portalocker>=2.0.0'
 
-.. GENERATED FROM PYTHON SOURCE LINES 24-28
+.. GENERATED FROM PYTHON SOURCE LINES 22-26
 
 Imports
 -------
 
 All imports used in the tutorial are declared at the top of the file.
 
-.. GENERATED FROM PYTHON SOURCE LINES 28-47
+.. GENERATED FROM PYTHON SOURCE LINES 26-44
 
 .. dropdown:: Code (Imports)
 
@@ -54,7 +52,6 @@ All imports used in the tutorial are declared at the top of the file.
 
         import ray
         import json
-        import pandas as pd
         from functools import partial
 
         import torch
@@ -76,17 +73,17 @@ All imports used in the tutorial are declared at the top of the file.
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 48-51
+.. GENERATED FROM PYTHON SOURCE LINES 45-48
 
 .. note::
   The following can be used to detect if **CUDA** devices are available on the current host. Therefore, this notebook will automatically adapt the parallel execution based on the ressources available locally. However, it will not be the case if many compute nodes are requested.
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 53-54
+.. GENERATED FROM PYTHON SOURCE LINES 50-51
 
 If GPU is available, this code will enabled the tutorial to use the GPU for pytorch operations.
 
-.. GENERATED FROM PYTHON SOURCE LINES 55-60
+.. GENERATED FROM PYTHON SOURCE LINES 52-57
 
 .. dropdown:: Code (Code to check if using CPU or GPU)
 
@@ -103,7 +100,7 @@ If GPU is available, this code will enabled the tutorial to use the GPU for pyto
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 61-66
+.. GENERATED FROM PYTHON SOURCE LINES 58-63
 
 The dataset
 -----------
@@ -111,7 +108,7 @@ The dataset
 The torchtext library provides a few raw dataset iterators, which yield the raw text strings. For example, the :code:`AG_NEWS` dataset iterators yield the raw data as a tuple of label and text. It has four labels (1 : World 2 : Sports 3 : Business 4 : Sci/Tec).
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 66-84
+.. GENERATED FROM PYTHON SOURCE LINES 63-81
 
 .. dropdown:: Code (Loading the data)
 
@@ -141,7 +138,7 @@ The torchtext library provides a few raw dataset iterators, which yield the raw 
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 85-93
+.. GENERATED FROM PYTHON SOURCE LINES 82-90
 
 Preprocessing pipelines and Batch generation
 --------------------------------------------
@@ -152,18 +149,18 @@ vocabulary.
 
[…]
