+++ joss/paper.md
@@ -73,7 +73,7 @@ While the feature matrix below provides a high-level overview, it necessarily si
 | Ensemble construction  | $\checkmark$ |              |              |
 | Visualization          | $\approx$    | $\checkmark$ |              |
 
-Table: The table bellow shows an overview of optimization features available in different hyperparameter optimization packages.
+Table: Overview of optimization features available in different packages.
 
 **Single-objective Optimization**  
 DeepHyper employs a surrogate model based on random forests to estimate $P(\text{Objective} \mid \text{Hyperparameters})$, similar to SMAC. However, DeepHyper's implementation is typically faster per query, especially when the number of evaluations exceeds 200. In contrast, Optuna uses the Tree-structured Parzen Estimator (TPE), which models $P(\text{Hyperparameters} \mid \text{Objective})$. TPE offers faster query times but can struggle with complex optimization landscapes and tends to be less effective in refining continuous hyperparameters.
@@ -107,6 +107,8 @@ Basic visualization tools are provided via `deephyper.analytics`. For more inter
 | Decentralized optimization | $\checkmark$ | $\checkmark$ |           |
 | Parallelization backends   | $\checkmark$ |              | $\approx$ |
 | Memory backends            | $\checkmark$ | $\checkmark$ | $\approx$ |
+
+Table: Overview of parallelization features available in different packages.
 
 The main difference between DeepHyper, Optuna and SMAC related to parallelization is that DeepHyper provides out-of-the-box parallelization software while Optuna leaves it to the user and SMAC limits itself to centralized parallelism with Dask.
 
