+++ joss/paper.md
@@ -38,7 +38,7 @@ Machine learning models are increasingly applied across scientific disciplines, 
 
 `DeepHyper` is a Python package for parallel hyperparameter optimization or neural architecture search. The project started in 2018 [@balaprakash2018deephyper] with a focus on making Bayesian optimization more efficient on high-performance computing clusters. It provides access to a variety of asynchronous parallel black-box optimization algorithms `deephyper.hpo`. The software offers a variety of parallel programming backends such as Asyncio, threading, processes, Ray, and MPI `deehyper.evaluator`. The hyperparameter optimization can be single or multi-objective, composed of mixed variables, using explicit or hidden constraints, and benefit from early-discarding strategies `deephyper.stopper`. Leveraging the results of hyperparameter optimization or neural architecture search it provides parallel ensemble algorithms `deephyper.ensemble` that can help improve accuracy or quantify disentangled predictive uncertainty. A diagram of our software architecture is shown in Figure 1.
 
-![Software Architecture](figures/deephyper-architecture.png){width=60%}
+![DeepHyper Software Architecture](figures/deephyper-architecture.png){width=50%}
 
 `DeepHyper` was designed to help research in the field of automated machine learning and also to be used out-of-the box in scientific projects where learning workflows are being developed.
 
@@ -59,7 +59,7 @@ These tools differ in scope and design: some are research-oriented (e.g., SMAC),
 
 ## Feature Comparison
 
-While the feature matrix below provides a high-level overview, it necessarily simplifies some nuanced implementation differences.
+While the feature matrix below provides a high-level overview, it necessarily simplifies some nuanced implementation differences. We use the $\checkmark$ symbol for available features, the $\approx$ symbol for incomplete features, and no symbol for missing features.
 
 ### Hyperparameter Optimization Capabilities
 
@@ -96,7 +96,7 @@ DeepHyper enables model ensembling from the pool of evaluated configurations, he
 **Visualization**  
 Basic visualization tools are provided via `deephyper.analytics`. For more interactive exploration, we recommend [SandDance](https://microsoft.github.io/SandDance/), a Visual Studio Code plugin. Figure 2 illustrates a 3D visualization of a Random Forest optimization, with `min_samples_split`, `min_weight_fraction_leaf`, and test accuracy as the x, y, and z axes (and color), respectively. The two plots compare configurations using `splitter="best"` (left) and `splitter="random"` (right). Such visualizations help identify the sensitivity of the objective to different hyperparameters.
 
-![Visualization of hyperparameter optimization results with SandDance](figures/sanddance_viz.png){width=80%,label="sanddance_viz"}
+![Visualization of hyperparameter optimization results with SandDance](figures/sanddance_viz.png){width=80%}
 
 ### Parallelization Capabilities
 
@@ -120,11 +120,12 @@ The main difference between DeepHyper, Optuna and SMAC related to parallelizatio
 
 **Parallelization backends**: DeepHyper's provides compatibility with several parallelization backends: AsyncIO functions, thread-based, process-based, Ray, and MPI. This allows to easily adapt to the context of the execution.
 
-**Memory backends**:: DeepHyper's provides compatibility with several shared-memory backends: local memory, server managed memory, MPI-based remote memory access, Ray.
+**Memory backends**: DeepHyper's provides compatibility with several shared-memory backends: local memory, server managed memory, MPI-based remote memory access, Ray.
 
 ## Black-box Optimization Benchmarks
 
[…]
