### 2025-04-25 deephyper/deephyper#313 @Deathn0t
Hello @jungtaekkim we just submitted an updated version of the manuscript that should address all your items in this issue. Let me know if it is good for you.
