### 2025-04-25 #7975 @jungtaekkim
Hi @jbytecode,

I would like to say that my reviews on manuscript and codebase are all reflected.

Thank you for your hard work and commitment to JOSS.
