### 2025-04-28 8a1457d
Remove unused input args in Search init and methods
