### 2025-05-05 f6590ee
Add csv_output argument for Search and CBO classes
