### 2025-05-05 642707b
Add argument to remove csv results file after it is created
