+++ src/deephyper/hpo/_cbo.py
@@ -301,7 +301,6 @@ class CBO(Search):
         log_dir: str = ".",
         verbose: int = 0,
         stopper: Optional[Stopper] = None,
-        csv_output: bool = True,
         surrogate_model="ET",
         surrogate_model_kwargs: Optional[SurrogateModelKwargs] = None,
         acq_func: str = "UCBd",
@@ -317,7 +316,7 @@ class CBO(Search):
         moo_scalarization_weight=None,
         objective_scaler="minmax",
     ):
-        super().__init__(problem, evaluator, random_state, log_dir, verbose, stopper, csv_output)
+        super().__init__(problem, evaluator, random_state, log_dir, verbose, stopper)
         # get the __init__ parameters
         self._init_params = locals()
 
