### 2025-05-06 93a68b6
Readd kwargs to CBO init
