### 2025-05-07 deephyper/deephyper#305 @y0z
@Deathn0t Thank you for the updates.

I confirmed the revised manuscript, and it is almost OK.
I leave two comments to be addressed. Once these are addressed, I think the revision will be complete.

1. The following is described in the multi-objective optimization section, but Optuna's default sampler for multi-objective optimization is NSGA-II, not MOTPE.

> While Optuna defaults to MoTPE,

https://optuna.readthedocs.io/en/stable/reference/generated/optuna.study.create_study.html#optuna.study.create_study

2. a typo in summary: accuray -> accuracy
