### 2025-05-07 #7975 @y0z
@jbytecode (cc @Deathn0t )

I've confirmed that all of my concerns have been addressed in the recent revision.
I have completed my review checklist.
