### 2025-05-07 deephyper/deephyper#305 @Deathn0t
After confirming that the items are resolved in https://github.com/openjournals/joss-reviews/issues/7975#issuecomment-2857363902
I am closing the issue.
