### 2025-05-07 #7975 @jbytecode
@Deathn0t - Could you please solve the invalid DOIs issue as our editorial bots reports?
