+++ joss/paper.bib
@@ -180,6 +180,7 @@
   pages     = {I–115–I–123},
   location  = {Atlanta, GA, USA},
   series    = {ICML'13},
+  url       = {https://dl.acm.org/doi/10.5555/3042817.3042832},
   doi       = {10.5555/3042817.3042832}
 }
 
@@ -207,6 +208,7 @@
   numpages  = {15},
   location  = {Vancouver, BC, Canada},
   series    = {NIPS '20},
+  url       = {https://dl.acm.org/doi/10.5555/3495724.3497531},
   doi       = {10.5555/3495724.3497531}
 }
 
@@ -218,7 +220,7 @@
   volume  = {25},
   number  = {120},
   pages   = {1--11},
-  url     = {http://jmlr.org/papers/v25/23-0537.html},
+  url     = {https://dl.acm.org/doi/10.5555/3722577.3722697},
   doi     = {10.5555/3722577.3722697}
 }
 
@@ -230,6 +232,6 @@
   volume  = {23},
   number  = {54},
   pages   = {1--9},
-  url     = {http://jmlr.org/papers/v23/21-0888.html},
+  url     = {https://dl.acm.org/doi/10.5555/3586589.3586643},
   doi     = {10.5555/3586589.3586643}
 }
