+++ joss/paper.bib
@@ -180,8 +180,7 @@
   pages     = {I–115–I–123},
   location  = {Atlanta, GA, USA},
   series    = {ICML'13},
-  url       = {https://dl.acm.org/doi/10.5555/3042817.3042832},
-  doi       = {10.5555/3042817.3042832}
+  url       = {https://dl.acm.org/doi/10.5555/3042817.3042832}
 }
 
 @article{hyperopt2015,
@@ -208,8 +207,7 @@
   numpages  = {15},
   location  = {Vancouver, BC, Canada},
   series    = {NIPS '20},
-  url       = {https://dl.acm.org/doi/10.5555/3495724.3497531},
-  doi       = {10.5555/3495724.3497531}
+  url       = {https://dl.acm.org/doi/10.5555/3495724.3497531}
 }
 
 @article{openbox2024,
@@ -220,8 +218,7 @@
   volume  = {25},
   number  = {120},
   pages   = {1--11},
-  url     = {https://dl.acm.org/doi/10.5555/3722577.3722697},
-  doi     = {10.5555/3722577.3722697}
+  url     = {https://dl.acm.org/doi/10.5555/3722577.3722697}
 }
 
 @article{smac32022,
@@ -232,6 +229,5 @@
   volume  = {23},
   number  = {54},
   pages   = {1--9},
-  url     = {https://dl.acm.org/doi/10.5555/3586589.3586643},
-  doi     = {10.5555/3586589.3586643}
+  url     = {https://dl.acm.org/doi/10.5555/3586589.3586643}
 }
