+++ joss/paper.bib
@@ -120,7 +120,8 @@
   title   = {Parallel Multi-Objective Hyperparameter Optimization with Uniform Normalization and Bounded Objectives},
   author  = {Égelé, Romain and Chang, Tyler and Sun, Yixuan and Vishwanath, Venkatram and Balaprakash, Prasanna},
   journal = {arXiv preprint arXiv:2309.14936},
-  year    = {2023}
+  year    = {2023},
+  doi     = {10.48550/arXiv.2309.14936}
 }
 
 @article{egele2024unreasonable,
