### 2025-05-08 #7975 @Deathn0t
@jbytecode I think we are good with DOIs now.
