### 2025-05-09 #7975 @jbytecode
@Deathn0t - I've just sent a pull request:

https://github.com/deephyper/deephyper/pull/321

Please review the changes carefully and merge them if you are agree with them.

Could you please review the post review checklist? We need a tagged release in the software repository and a Zenodo archive which is created using the release. The aim is to store the latest reviewed form the project.  Please ping me when the stuff is ready.
