### 2025-05-09 165377a
Fixes on Paper and Bibtex (#321)
