+++ joss/paper.bib
@@ -70,7 +70,7 @@
 @inproceedings{dorier2022transferlearning,
   author    = {M. Dorier and R. Égelé and P. Balaprakash and J. Koo and S. Madireddy and S. Ramesh and A. D. Malony and R. Ross},
   booktitle = {2022 IEEE International Conference on Cluster Computing (CLUSTER)},
-  title     = {HPC Storage Service Autotuning Using Variational- Autoencoder -Guided Asynchronous Bayesian Optimization},
+  title     = {HPC Storage Service Autotuning Using Variational- Autoencoder -Guided Asynchronous {B}ayesian Optimization},
   year      = {2022},
   volume    = {},
   issn      = {},
@@ -83,7 +83,7 @@
 }
 
 @inproceedings{egele2023asynchronous,
-  title        = {Asynchronous Decentralized Bayesian Optimization for Large Scale Hyperparameter Optimization},
+  title        = {Asynchronous Decentralized {B}ayesian Optimization for Large Scale Hyperparameter Optimization},
   author       = {Egel{\'e}, Romain and Guyon, Isabelle and Vishwanath, Venkatram and Balaprakash, Prasanna},
   booktitle    = {2023 IEEE 19th International Conference on e-Science (e-Science)},
   pages        = {1--10},
@@ -167,7 +167,7 @@
   booktitle = {Proceedings of the 25th ACM SIGKDD International Conference on Knowledge Discovery \& Data Mining},
   pages     = {2623–2631},
   numpages  = {9},
-  keywords  = {machine learning system, hyperparameter optimization, black-box optimization, Bayesian optimization},
+  keywords  = {machine learning system, hyperparameter optimization, black-box optimization, {B}ayesian optimization},
   location  = {Anchorage, AK, USA},
   series    = {KDD '19}
 }
@@ -185,7 +185,7 @@
 }
 
 @article{hyperopt2015,
-  title     = {Hyperopt: a python library for model selection and hyperparameter optimization},
+  title     = {Hyperopt: a {P}ython library for model selection and hyperparameter optimization},
   author    = {Bergstra, James and Komer, Brent and Eliasmith, Chris and Yamins, Dan and Cox, David D},
   journal   = {Computational Science \& Discovery},
   volume    = {8},
@@ -198,7 +198,7 @@
 
 @inproceedings{botorch2020,
   author    = {Balandat, Maximilian and Karrer, Brian and Jiang, Daniel R. and Daulton, Samuel and Letham, Benjamin and Wilson, Andrew Gordon and Bakshy, Eytan},
-  title     = {BOTORCH: a framework for efficient monte-carlo Bayesian optimization},
+  title     = {BOTORCH: a framework for efficient monte-carlo {B}ayesian optimization},
   year      = {2020},
   isbn      = {9781713829546},
   publisher = {Curran Associates Inc.},
@@ -213,7 +213,7 @@
 
 @article{openbox2024,
   author  = {Huaijun Jiang and Yu Shen and Yang Li and Beicheng Xu and Sixian Du and Wentao Zhang and Ce Zhang and Bin Cui},
-  title   = {OpenBox: A Python Toolkit for Generalized Black-box Optimization},
+  title   = {OpenBox: A {P}ython Toolkit for Generalized Black-box Optimization},
   journal = {Journal of Machine Learning Research},
   year    = {2024},
   volume  = {25},
@@ -224,7 +224,7 @@
 
 @article{smac32022,
   author  = {Marius Lindauer and Katharina Eggensperger and Matthias Feurer and AndrÃ© Biedenkapp and Difan Deng and Carolin Benjamins and Tim Ruhkopf and RenÃ© Sass and Frank Hutter},
[…]
