### 2025-05-12 ceaf1a7
Integration of JOSS submition (#322)
