+++ joss/paper.md
@@ -1,0 +1,172 @@
+---
+title: 'DeepHyper: A Python Package for Massively Parallel Hyperparameter Optimization in Machine Learning'
+tags:
+  - Python
+  - machine learning
+  - hyperparameter optimization
+  - multi-fidelity
+  - neural architecture search
+  - ensemble
+  - high-performance computing
+authors:
+  - name: Romain Egele
+    orcid: 0000-0002-8992-8192
+    equal-contrib: true
+    affiliation: "1"
+    corresponding: true
+  - name: Prasanna Balaprakash
+    orcid: 0000-0002-0292-5715
+    equal-contrib: true
+    affiliation: "1"
+  - name: Gavin M. Wiggins
+    affiliation: 1
+  - name: Brett Eiffert
+    affiliation: 1
+affiliations:
+  - name: Oak Ridge National Laboratory, Oak Ridge, TN, United States
+    index: 1
+    ror: 01qz5mb56
+date: 3 February 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+Machine learning models are increasingly applied across scientific disciplines, yet their effectiveness often hinges on heuristic decisions—such as data transformations, training strategies, and model architectures—that are not learned by the models themselves. Automating the selection of these heuristics and analyzing their sensitivity is crucial for building robust and efficient learning workflows. `DeepHyper` addresses this challenge by democratizing hyperparameter optimization, providing accessible tools to streamline and enhance machine learning workflows from a laptop to the largest supercomputer in the world. Building on top of hyperparameter optimization, it unlocks new capabilities around ensembles of models for improved accuracy and uncertainty quantification. All of these organized around efficient parallel computing.
+
+# Statement of need
+
+`DeepHyper` is a Python package for parallel hyperparameter optimization or neural architecture search. The project started in 2018 [@balaprakash2018deephyper] with a focus on making Bayesian optimization more efficient on high-performance computing clusters. It provides access to a variety of asynchronous parallel black-box optimization algorithms via `deephyper.hpo`. The software offers a variety of parallel programming backends such as Asyncio, threading, processes, Ray, and MPI `deehyper.evaluator`. The hyperparameter optimization can be single or multi-objective, composed of mixed variables, using explicit or hidden constraints, and benefit from early-discarding strategies via `deephyper.stopper`. Leveraging the results of hyperparameter optimization or neural architecture search it provides parallel ensemble algorithms via `deephyper.ensemble` that can help improve accuracy or quantify disentangled predictive uncertainty. A diagram of our software architecture is shown in Figure 1.
+
+![DeepHyper Software Architecture](figures/deephyper-architecture.png){width=50%}
+
+`DeepHyper` was designed to help research in the field of automated machine learning and also to be used out-of-the box in scientific projects where learning workflows are being developed.
+
+# Related Work
+
+Numerous software packages now exist for hyperparameter optimization (HPO), including:
+
+- BoTorch (and Ax platform) [@botorch2020], ([doc.](https://botorch.org))
+- HyperMapper [@hypermapper2019], ([doc.](https://github.com/luinardi/hypermapper))
+- Hyperopt [@hyperopt2013;@hyperopt2015], ([doc.](http://hyperopt.github.io/hyperopt/))
+- OpenBox [@openbox2024], ([doc.](https://open-box.readthedocs.io))
+- Optuna [@optuna2019], ([doc.](https://optuna.readthedocs.io))
+- SMAC3 [@smac32022], ([doc.](https://automl.github.io/SMAC3))
+
[…]
