### 2025-05-12 #7975 @Deathn0t
@jbytecode I created the tag `joss-1` https://github.com/deephyper/deephyper/tree/joss-1

in the post-review check list I don't know why but the LICENSE used in the PDF does not correspond to the LICENSE found by the bot in the repository.
