### 2025-05-12 #7975 @Deathn0t
@jbytecode  - The version is `0.10.1` the Zenodo link: https://zenodo.org/records/15387374
