### 2025-05-12 #7975 @jbytecode
@Deathn0t - Looks almost good to me, could you please have one more proof-read and fix anything if exists. Please ping when when you are done. Thank you in advance.
