### 2025-05-12 af74129
feat (cbo): adding parameter outliers_iqr_factor to CBO
