+++ src/deephyper/hpo/_cbo.py
@@ -67,6 +67,7 @@ class AcqOptimizerKwargs(BaseModel):
     acq_optimizer_freq: Optional[int] = 1
     n_jobs: Optional[int] = 1
     n_restarts_optimizer: Optional[int] = 1
+    outliers_iqr_factor: Optional[float] = 1.5
     # Genetic algorithms parameters
     ga_pop_size: Optional[int] = 100
     ga_xtol: Optional[float] = 1e-8
@@ -334,8 +335,12 @@ class CBO(Search):
             "DUMMY",
         ]
 
-        surrogate_model_kwargs = {} if surrogate_model_kwargs is None else surrogate_model_kwargs
-        self._surrogate_model_kwargs = SurrogateModelKwargs(**surrogate_model_kwargs).model_dump()
+        surrogate_model_kwargs = (
+            {} if surrogate_model_kwargs is None else surrogate_model_kwargs
+        )
+        self._surrogate_model_kwargs = SurrogateModelKwargs(
+            **surrogate_model_kwargs
+        ).model_dump()
 
         base_estimator_scheduler = self._surrogate_model_kwargs.pop("scheduler")
 
@@ -373,8 +378,12 @@ class CBO(Search):
         acq_func_kwargs = {} if acq_func_kwargs is None else acq_func_kwargs
         self._acq_func_kwargs = AcqFuncKwargs(**acq_func_kwargs).model_dump()
 
-        acq_optimizer_kwargs = {} if acq_optimizer_kwargs is None else acq_optimizer_kwargs
-        self._acq_optimizer_kwargs = AcqOptimizerKwargs(**acq_optimizer_kwargs).model_dump()
+        acq_optimizer_kwargs = (
+            {} if acq_optimizer_kwargs is None else acq_optimizer_kwargs
+        )
+        self._acq_optimizer_kwargs = AcqOptimizerKwargs(
+            **acq_optimizer_kwargs
+        ).model_dump()
 
         # Initialize lower bounds for objectives
         if moo_lower_bounds is None:
@@ -383,7 +392,8 @@ class CBO(Search):
             [isinstance(lbi, numbers.Number) or lbi is None for lbi in moo_lower_bounds]
         ):
             self._moo_upper_bounds = [
-                -lbi if isinstance(lbi, numbers.Number) else None for lbi in moo_lower_bounds
+                -lbi if isinstance(lbi, numbers.Number) else None
+                for lbi in moo_lower_bounds
             ]
         else:
             raise ValueError(
@@ -466,12 +476,13 @@ class CBO(Search):
             moo_scalarization_strategy=self._moo_scalarization_strategy,
             moo_scalarization_weight=self._moo_scalarization_weight,
             objective_scaler=objective_scaler,
+            outliers_iqr_factor=self._acq_optimizer_kwargs.get("outliers_iqr_factor"),
         )
 
         # Scheduler policy
         scheduler = self._acq_func_kwargs["scheduler"]
         scheduler = {"type": "bandit"} if scheduler is None else scheduler
-        if scheduler["delay"] == "n-initial-points":
+        if scheduler is not None and scheduler["delay"] == "n-initial-points":
             scheduler["delay"] = self._n_initial_points
 
         self.scheduler = None
@@ -513,7 +524,9 @@ class CBO(Search):
                 }
                 scheduler_func = scheduler_bandit
 
-            eta_0 = np.array([self._acq_func_kwargs["kappa"], self._acq_func_kwargs["xi"]])
+            eta_0 = np.array(
+                [self._acq_func_kwargs["kappa"], self._acq_func_kwargs["xi"]]
+            )
             self.scheduler = functools.partial(
                 scheduler_func,
                 eta_0=eta_0,
[…]
