### 2025-05-12 #7975 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
