### 2025-05-12 4e8dccc
queued_evaluator_mpi not working in sphinx gallery
