+++ src/deephyper/hpo/_cbo.py
@@ -335,12 +335,8 @@ class CBO(Search):
             "DUMMY",
         ]
 
-        surrogate_model_kwargs = (
-            {} if surrogate_model_kwargs is None else surrogate_model_kwargs
-        )
-        self._surrogate_model_kwargs = SurrogateModelKwargs(
-            **surrogate_model_kwargs
-        ).model_dump()
+        surrogate_model_kwargs = {} if surrogate_model_kwargs is None else surrogate_model_kwargs
+        self._surrogate_model_kwargs = SurrogateModelKwargs(**surrogate_model_kwargs).model_dump()
 
         base_estimator_scheduler = self._surrogate_model_kwargs.pop("scheduler")
 
@@ -378,12 +374,8 @@ class CBO(Search):
         acq_func_kwargs = {} if acq_func_kwargs is None else acq_func_kwargs
         self._acq_func_kwargs = AcqFuncKwargs(**acq_func_kwargs).model_dump()
 
-        acq_optimizer_kwargs = (
-            {} if acq_optimizer_kwargs is None else acq_optimizer_kwargs
-        )
-        self._acq_optimizer_kwargs = AcqOptimizerKwargs(
-            **acq_optimizer_kwargs
-        ).model_dump()
+        acq_optimizer_kwargs = {} if acq_optimizer_kwargs is None else acq_optimizer_kwargs
+        self._acq_optimizer_kwargs = AcqOptimizerKwargs(**acq_optimizer_kwargs).model_dump()
 
         # Initialize lower bounds for objectives
         if moo_lower_bounds is None:
@@ -392,8 +384,7 @@ class CBO(Search):
             [isinstance(lbi, numbers.Number) or lbi is None for lbi in moo_lower_bounds]
         ):
             self._moo_upper_bounds = [
-                -lbi if isinstance(lbi, numbers.Number) else None
-                for lbi in moo_lower_bounds
+                -lbi if isinstance(lbi, numbers.Number) else None for lbi in moo_lower_bounds
             ]
         else:
             raise ValueError(
@@ -524,9 +515,7 @@ class CBO(Search):
                 }
                 scheduler_func = scheduler_bandit
 
-            eta_0 = np.array(
-                [self._acq_func_kwargs["kappa"], self._acq_func_kwargs["xi"]]
-            )
+            eta_0 = np.array([self._acq_func_kwargs["kappa"], self._acq_func_kwargs["xi"]])
             self.scheduler = functools.partial(
                 scheduler_func,
                 eta_0=eta_0,
@@ -539,9 +528,7 @@ class CBO(Search):
         elif callable(scheduler):
             self.scheduler = functools.partial(
                 scheduler,
-                eta_0=np.array(
-                    [self._acq_func_kwargs["kappa"], self._acq_func_kwargs["xi"]]
-                ),
+                eta_0=np.array([self._acq_func_kwargs["kappa"], self._acq_func_kwargs["xi"]]),
             )
             logging.info(f"Set up scheduler '{scheduler}'")
 
@@ -557,9 +544,7 @@ class CBO(Search):
         if self.scheduler is not None:
             kappa, xi = self.scheduler(i)
             values = {"kappa": float(kappa), "xi": float(xi)}
-            logging.info(
-                f"Updated exploration-exploitation policy with {values} from scheduler"
-            )
+            logging.info(f"Updated exploration-exploitation policy with {values} from scheduler")
             self._opt.acq_func_kwargs.update(values)
 
     def _ask(self, n: int = 1) -> List[Dict]:
@@ -601,10 +586,7 @@ class CBO(Search):
             elif (type(obj) is str and "F" == obj[0]) or any(
[…]
