### 2025-05-12 809e001
feat (CBO): adding outliers_iqr_factor to CBO (#324)
* feat (cbo): adding parameter outliers_iqr_factor to CBO

* format with ruff
