+++ src/deephyper/hpo/_search.py
@@ -1,6 +1,7 @@
 import abc
 import asyncio
 import copy
+import csv
 import json
 import logging
 import os
@@ -74,6 +75,7 @@ class Search(abc.ABC):
         log_dir=".",
         verbose=0,
         stopper=None,
+        csv_output=True,
     ):
         # TODO: stopper should be an argument passed here... check CBO and generalize
         # get the __init__ parameters
@@ -129,8 +131,8 @@ class Search(abc.ABC):
         self.gather_batch_size = 1
 
         self._evaluator._stopper = stopper
-
         self.stopped = False
+        self.csv_output = csv_output
 
     def check_evaluator(self, evaluator):
         if not (isinstance(evaluator, Evaluator)):
@@ -190,51 +192,30 @@ class Search(abc.ABC):
                 raise ValueError("'timeout' should be > 0!")
 
     def search(
-        self,
-        max_evals: int = -1,
-        timeout: int = None,
-        max_evals_strict: bool = False,
-        remove_csv: bool = False,
-    ):
+        self, max_evals=-1, timeout=0, max_evals_strict=False, csv_output=True
+    ) -> pd.DataFrame:
         """Execute the search algorithm.
 
         Args:
-            max_evals (int, optional): The maximum number of evaluations of the run function to
-                perform before stopping the search. Defaults to ``-1``, will run indefinitely.
-            timeout (int, optional): The time budget (in seconds) of the search before stopping.
-                Defaults to ``None``, will not impose a time budget.
-            max_evals_strict (bool, optional): If ``True`` the search will not spawn more than
-                ``max_evals`` jobs. Defaults to ``False``.
-            remove_csv: Remove the CSV results file.
+            max_evals: here
 
         Returns:
-            DataFrame: A pandas DataFrame containing the evaluations performed or ``None`` if the
-                search could not evaluate any configuration.
-
-                This DataFrame contains the following columns:
-                - ``p:HYPERPARAMETER_NAME``: for each hyperparameter of the problem.
-                - ``objective``: for single objective optimization.
-                - ``objective_0``, ``objective_1``, ...: for multi-objective optimization.
-                - ``job_id``: the identifier of the job.
-                - ``job_status``: the status of the job at the end of the search.
-                - ``m:METADATA_NAME``: for each metadata of the problem. Some metadata are always
-                    present like ``m:timestamp_submit`` and ``m:timestamp_gather`` which are the
-                    timestamps of the submission and gathering of the job.
+            here
         """
         self.stopped = False
-        self._check_timeout(timeout)
+        self.csv_output = csv_output
 
         if max_evals_strict:
             # TODO: should be replaced by a property with a setter?
             self._evaluator.set_maximum_num_jobs_submitted(max_evals)
 
-        # save the search call arguments for the context
+        # Save the search call arguments for the context
         self._call_args.append({"timeout": timeout, "max_evals": max_evals})
 
-        # save the context in the log folder
+        # Save the context in the log folder
         self.dump_context()
 
-        # init tqdm callback
+        # Setup the tqdm callback
         if max_evals > 1:
             for cb in self._evaluator._callbacks:
                 if isinstance(cb, TqdmCallback):
@@ -243,15 +224,13 @@ class Search(abc.ABC):
         wait_all_running_jobs = True
 
         try:
-            if np.isscalar(timeout) and timeout > 0:
+            if timeout > 0:
                 self._evaluator.timeout = timeout
             self._search(max_evals, max_evals_strict)
         except MaximumJobsSpawnReached:
             self.stopped = True
-            logging.warning(
[…]
