### 2025-05-13 a0f39f3
chore (action): adding workflow_dispatch to checks and tests
