### 2025-05-14 74d319f
Refactor test__evaluator to use dump_job_results method
