+++ tests/evaluator/test__evaluator.py
@@ -67,7 +67,7 @@ def test_run_function_standards(tmp_path):
 
     evaluator.submit(configs)
     evaluator.gather(type="ALL")
-    evaluator.dump_jobs_done_to_csv(log_dir=tmp_path)
+    evaluator.dump_job_results(log_dir=tmp_path, filename="results.csv", csv_output=True)
     results = pd.read_csv(os.path.join(tmp_path, "results.csv"))
     assert all(
         results.columns
@@ -95,7 +95,7 @@ def test_run_function_standards(tmp_path):
     evaluator._job_class = HPOJob
     evaluator.submit(configs)
     evaluator.gather(type="ALL")
-    evaluator.dump_jobs_done_to_csv(log_dir=tmp_path)
+    evaluator.dump_job_results(log_dir=tmp_path, filename="results.csv", csv_output=True)
     results = pd.read_csv(os.path.join(tmp_path, "results.csv")).sort_values(by="job_id")
     assert all(
         results.columns
@@ -120,7 +120,7 @@ def test_run_function_standards(tmp_path):
     evaluator._job_class = HPOJob
     evaluator.submit(configs)
     evaluator.gather(type="ALL")
-    evaluator.dump_jobs_done_to_csv(log_dir=tmp_path)
+    evaluator.dump_job_results(log_dir=tmp_path, filename="results.csv", csv_output=True)
     results = pd.read_csv(os.path.join(tmp_path, "results.csv"))
     assert all(
         results.columns
@@ -148,7 +148,7 @@ def test_run_function_standards(tmp_path):
     evaluator._job_class = HPOJob
     evaluator.submit(configs)
     evaluator.gather(type="ALL")
-    evaluator.dump_jobs_done_to_csv(log_dir=tmp_path)
+    evaluator.dump_job_results(log_dir=tmp_path, filename="results.csv", csv_output=True)
     results = pd.read_csv(os.path.join(tmp_path, "results.csv"))
     assert list(sorted(results.columns)) == list(
         sorted(
@@ -181,7 +181,7 @@ def test_run_function_standards(tmp_path):
     evaluator._job_class = HPOJob
     evaluator.submit(configs)
     evaluator.gather(type="ALL")
-    evaluator.dump_jobs_done_to_csv(log_dir=tmp_path)
+    evaluator.dump_job_results(log_dir=tmp_path, filename="results.csv", csv_output=True)
     results = pd.read_csv(os.path.join(tmp_path, "results.csv"))
     assert list(sorted(results.columns)) == list(
         sorted(
@@ -216,7 +216,7 @@ def test_run_function_standards(tmp_path):
     evaluator._job_class = HPOJob
     evaluator.submit(configs)
     evaluator.gather(type="ALL")
-    evaluator.dump_jobs_done_to_csv(log_dir=tmp_path)
+    evaluator.dump_job_results(log_dir=tmp_path, filename="results.csv", csv_output=True)
     results = pd.read_csv(os.path.join(tmp_path, "results.csv"))
     assert list(sorted(results.columns)) == list(
         sorted(
@@ -255,7 +255,7 @@ def test_run_function_standards(tmp_path):
     evaluator._job_class = HPOJob
     evaluator.submit(configs)
     evaluator.gather(type="ALL")
-    evaluator.dump_jobs_done_to_csv(log_dir=tmp_path)
+    evaluator.dump_job_results(log_dir=tmp_path, filename="results.csv", csv_output=True)
     results = pd.read_csv(os.path.join(tmp_path, "results.csv"))
     assert list(sorted(results.columns)) == list(
         sorted(
[…]
