### 2025-05-14 c86dbcd
chore (search): refactoring management of results
