+++ src/deephyper/hpo/_search.py
@@ -1,6 +1,7 @@
 import abc
 import asyncio
 import copy
+import csv
 import json
 import logging
 import os
@@ -45,6 +46,154 @@ def get_init_params_as_json(obj):
                 except Exception:
                     params[k] = "NA"
     return params
+
+
+class SearchHistory:
+    """Represents the history of a search."""
+
+    def __init__(self, problem):
+        self.problem = problem
+        self.num_objective = None
+        self.jobs = []
+        self._csv_cursor = 0
+        self._csv_columns = None
+
+    def __getitem__(self, idx) -> HPOJob:
+        return self.jobs[idx]
+
+    def append(self, job: HPOJob):
+        self.jobs.append(job)
+
+    def extend(self, jobs: List[HPOJob]):
+        self.jobs.extends(jobs)
+
+    def _to_dict(self, jobs: List[HPOJob]) -> List[dict]:
+        jobs_as_dict = []
+        for job in jobs:
+            result = copy.deepcopy(job.args)
+
+            # add prefix for all keys found in "args"
+            result = {f"p:{k}": v for k, v in result.items()}
+
+            # when the returned value of the run-function is a dict we flatten it to add in csv
+            result["objective"] = job.objective
+
+            # when the objective is a tuple (multi-objective) we create 1 column per tuple-element
+            if isinstance(result["objective"], tuple) or isinstance(result["objective"], list):
+                obj = result.pop("objective")
+
+                if self.num_objective is None:
+                    self.num_objective = len(obj)
+
+                for i, objval in enumerate(obj):
+                    result[f"objective_{i}"] = objval
+            else:
+                if self.num_objective is None:
+                    self.num_objective = 1
+
+                if self.num_objective > 1:
+                    obj = result.pop("objective")
+                    for i in range(self.num_objective):
+                        result[f"objective_{i}"] = obj
+
+            # Add job.id
+            result["job_id"] = int(job.id.split(".")[1])
+
+            # Add job.status
+            result["job_status"] = job.status.name
+
+            # Profiling and other
+            # metadata keys starting with "_" are not saved (considered as internal/private)
+            metadata = {f"m:{k}": v for k, v in job.metadata.items() if k[0] != "_"}
+            result.update(metadata)
+
+            jobs_as_dict.append(result)
+
+        return jobs_as_dict
+
+    def to_dataframe(self) -> pd.Dataframe:
+        df = pd.DataFrame(self._to_dict(self.jobs))
+        return df
+
+    def to_csv(self, path: str, partial: bool = False, flush: bool = False):
+        """Write results to CSV file.
+
+        Args:
+            path (str): _description_
+            partial (bool, optional): _description_. Defaults to False.
+            flush (bool, optional): ...
+        """
+        if not partial:
+            df = self.to_dataframe()
+            df.to_csv(path, index=False)
+            return
+
+        resultsList = []
+
+        for job in self.jobs[self._csv_cursor :]:
+            result = copy.deepcopy(job.args)
+
+            # add prefix for all keys found in "args"
+            result = {f"p:{k}": v for k, v in result.items()}
+
+            # when the returned value of the run-function is a dict we flatten it to add in csv
+            result["objective"] = job.objective
+
+            # when the objective is a tuple (multi-objective) we create 1 column per tuple-element
+            if isinstance(result["objective"], tuple) or isinstance(result["objective"], list):
+                obj = result.pop("objective")
+
+                if self.num_objective is None:
+                    self.num_objective = len(obj)
+
+                for i, objval in enumerate(obj):
+                    result[f"objective_{i}"] = objval
+            else:
+                if self.num_objective is None:
[…]
