### 2025-05-14 eb5b22a
chore (search): search now using the SearchHisory class also for the pareto-efficient attribute
