+++ src/deephyper/hpo/_search.py
@@ -55,8 +55,12 @@ class SearchHistory:
         self.problem = problem
         self.num_objective = None
         self.jobs = []
+        self.pareto_efficient = []
         self._csv_cursor = 0
         self._csv_columns = None
+
+    def __len__(self):
+        return len(self.jobs)
 
     def __getitem__(self, idx) -> HPOJob:
         return self.jobs[idx]
@@ -65,7 +69,7 @@ class SearchHistory:
         self.jobs.append(job)
 
     def extend(self, jobs: List[HPOJob]):
-        self.jobs.extends(jobs)
+        self.jobs.extend(jobs)
 
     def _to_dict(self, jobs: List[HPOJob]) -> List[dict]:
         jobs_as_dict = []
@@ -107,11 +111,14 @@ class SearchHistory:
             metadata = {f"m:{k}": v for k, v in job.metadata.items() if k[0] != "_"}
             result.update(metadata)
 
+            if hasattr(job, "pareto_efficient"):
+                result["pareto_efficient"] = job.pareto_efficient
+
             jobs_as_dict.append(result)
 
         return jobs_as_dict
 
-    def to_dataframe(self) -> pd.Dataframe:
+    def to_dataframe(self) -> pd.DataFrame:
         df = pd.DataFrame(self._to_dict(self.jobs))
         return df
 
@@ -119,56 +126,27 @@ class SearchHistory:
         """Write results to CSV file.
 
         Args:
-            path (str): _description_
-            partial (bool, optional): _description_. Defaults to False.
-            flush (bool, optional): ...
+            path (str):
+
+                Path of the CSV file.
+
+            partial (bool, optional):
+
+                A boolean that indicates if the CSV write is only writing the last non-writen
+                chunk. Defaults to ``False``.
+
+            flush (bool, optional):
+
+                A boolean that indicates if the CSV write in the case where ``partial=True`` should
+                be flushed anyway. Otherwise if ``False`` it will not write to disk until there is
+                a successful job. Defaults to ``False``.
         """
         if not partial:
             df = self.to_dataframe()
             df.to_csv(path, index=False)
             return
 
-        resultsList = []
-
-        for job in self.jobs[self._csv_cursor :]:
-            result = copy.deepcopy(job.args)
-
-            # add prefix for all keys found in "args"
-            result = {f"p:{k}": v for k, v in result.items()}
-
-            # when the returned value of the run-function is a dict we flatten it to add in csv
-            result["objective"] = job.objective
-
-            # when the objective is a tuple (multi-objective) we create 1 column per tuple-element
-            if isinstance(result["objective"], tuple) or isinstance(result["objective"], list):
-                obj = result.pop("objective")
-
-                if self.num_objective is None:
-                    self.num_objective = len(obj)
-
-                for i, objval in enumerate(obj):
-                    result[f"objective_{i}"] = objval
-            else:
-                if self.num_objective is None:
-                    self.num_objective = 1
-
-                if self.num_objective > 1:
-                    obj = result.pop("objective")
-                    for i in range(self.num_objective):
-                        result[f"objective_{i}"] = obj
-
-            # Add job.id
-            result["job_id"] = int(job.id.split(".")[1])
-
-            # Add job.status
-            result["job_status"] = job.status.name
-
-            # Profiling and other
-            # methdata keys starting with "_" are not saved (considered as internal)
-            metadata = {f"m:{k}": v for k, v in job.metadata.items() if k[0] != "_"}
-            result.update(metadata)
-
-            resultsList.append(result)
+        resultsList = self._to_dict(self.jobs[self._csv_cursor :])
 
         if len(resultsList) > 0:
[…]
