### 2025-05-15 bf57cc4
replaced processpool with loky, added more example descriptions
