+++ docs/examples/examples_parallelism/plot_queued_evaluator_mpi.rst
@@ -1,0 +1,215 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "examples/examples_parallelism/plot_queued_evaluator_mpi.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_examples_examples_parallelism_plot_queued_evaluator_mpi.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_examples_examples_parallelism_plot_queued_evaluator_mpi.py:
+
+
+Queued Evaluator with MPI
+=========================
+
+**Author(s)**: Romain Egele, Brett Eiffert.
+
+In this example, you will learn how to use a queued ProcessPoolEvaluator with an mpi run function.
+
+.. GENERATED FROM PYTHON SOURCE LINES 11-25
+
+.. dropdown:: Code (Import statements)
+
+    .. code-block:: Python
+
+
+        import subprocess
+
+        from deephyper.evaluator import (
+            LokyEvaluator,
+            parse_subprocess_result,
+            profile,
+            queued,
+        )
+        from deephyper.evaluator.callback import TqdmCallback
+        from deephyper.hpo import CBO, HpProblem
+
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 26-30
+
+Run function using MPI
+----------------
+
+Used by the main function to fork and manage mpi processes within the evaluator.
+
+.. GENERATED FROM PYTHON SOURCE LINES 30-43
+
+.. code-block:: Python
+
+    @profile
+    def run_mpi_exe(job, dequed=None):
+        x = job.parameters["x"]
+
+        # The format of the print "DH-OUTPUT:(.+)\n" is strict if you use parse_suprocess_result
+        command = f"mpirun -np {len(dequed)} echo DH-OUTPUT:{x}\n"
+        completed_process = subprocess.run(command.split(), capture_output=True)
+        objective = parse_subprocess_result(completed_process)
+
+        # In the results.csv a new `m:dequed` will show the passed dequed values.
+        return {"objective": objective, "metadata": {}}
+
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 44-48
+
+Setup
+-----
+
+The problem is defined with basic hyperparameters for a straightforward example.
+
+.. GENERATED FROM PYTHON SOURCE LINES 48-52
+
+.. code-block:: Python
+
+
+    problem = HpProblem()
+    problem.add_hyperparameter((0.0, 10.0), "x")
+
+
+
+
+
+.. rst-class:: sphx-glr-script-out
+
+ .. code-block:: none
+
+
+    UniformFloatHyperparameter(name='x', default_value=5.0, meta=None, size=inf, lower=0.0, upper=10.0, log=False)
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 53-56
+
+Variables used for selecting the number of workers to execute the pool of mpi workers.
+These are defined for show and can be run in a multi-node setup or on a single node or local machine.
+Number of processes spawned = num_nodes / num_nodes_per_task
+
[…]
