### 2025-05-16 162a76c
chore (search): refactoring detection of rows with failures
