+++ tests/hpo/test__cbo.py
@@ -536,7 +536,7 @@ def test_cbo_checkpoint_restart_with_failures(tmp_path):
         surrogate_model_kwargs=SEARCH_KWARGS_DEFAULTS["surrogate_model_kwargs"],
         **search_kwargs,
     )
-
+    
     search_b.fit_surrogate(results_a)
     new_results_b = search_b.search(20)
     assert len(new_results_b) == 20
@@ -739,4 +739,6 @@ def test_cbo_fit_generative_model(tmp_path):
 if __name__ == "__main__":
     # test_sample_types(".")
     # test_gp(".")
-    test_cbo_categorical_variable(".")
+    # test_cbo_categorical_variable(".")
+    # test_cbo_checkpoint_restart_moo_with_failures(".")
+    test_cbo_checkpoint_restart_with_failures(".")
