### 2025-05-16 7c8acb5
chore (tests): adding a test to run a search without dumping csv file
