+++ tests/stopper/test__sha_stopper.py
@@ -1,7 +1,6 @@
 import time
 
 import numpy as np
-import pandas as pd
 import pytest
 
 from deephyper.evaluator import RunningJob
