### 2025-05-16 562eefe
chore (refactor): improved implementation of conversion to list of dict in history
