### 2025-05-16 d2328e8
Add queued evaluator mpi tutorial (#327)
* queued_evaluator_mpi not working in sphinx gallery

* replaced processpool with loky, added more example descriptions
