### 2025-05-16 4e5cf47
chore (refactor): adding csvlogger callback for evaluator, adding on_gather callback for callback interface, cleaning up evaluator code, adapting code in search
