+++ src/deephyper/evaluator/_evaluator.py
@@ -1,7 +1,5 @@
 import abc
 import asyncio
-import copy
-import csv
 import importlib
 import json
 import logging
@@ -87,7 +85,6 @@ class Evaluator(abc.ABC):
         self._tasks_running = []  # List of AsyncIO Task objects currently running.
         self._tasks_done = []  # Temp list to hold completed tasks from asyncio.
         self._tasks_pending = []  # Temp list to hold pending tasks from asyncio.
-        self.jobs_done = []  # List used to store all jobs completed by the evaluator.
         self.job_id_submitted = []  # List of jobs'id submitted by the evaluator.
         self.job_id_gathered = []  # List of jobs'id gathered by the evaluator.
         self.timestamp = time.time()  # Recorded time of when this evaluator interface was created.
@@ -405,6 +402,10 @@ class Evaluator(abc.ABC):
         # Access storage to return results from other processes
         other_results = self.gather_other_jobs_done()
 
+        # call callbacks
+        for cb in self._callbacks:
+            cb.on_gather(local_results, other_results)
+
         if len(other_results) == 0:
             logging.info(f"gather done - {len(local_results)} job(s)")
 
@@ -445,7 +446,6 @@ class Evaluator(abc.ABC):
                     job.metadata.update(job_data["metadata"])
                     job.set_output(job_data["out"])
                     self.job_id_gathered.append(job_id)
-                    self.jobs_done.append(job)
                     other_results.append(job)
 
                     for cb in self._callbacks:
@@ -485,7 +485,6 @@ class Evaluator(abc.ABC):
             job = task.result()
             self._on_done(job)
             local_results.append(job)
-            self.jobs_done.append(job)
             self._tasks_running.remove(task)
             self.job_id_gathered.append(job.id)
             self.job_id_submitted.remove(job.id)
@@ -495,18 +494,16 @@ class Evaluator(abc.ABC):
 
         return local_results
 
-    async def _await_cancelling_of_running_tasks(self):
-        self._tasks_done, self._tasks_pending = await asyncio.wait(
-            self._tasks_running,
-            return_when="ALL_COMPLETED",
-        )
-        # self._tasks_running = []
-
-    def close(self):
+    def close(self) -> List[Job]:
         logging.info(f"Closing {type(self).__name__}")
-
+        jobs = []
         if self.loop is None:
-            return
+            # calling callbacks
+            for cb in self._callbacks:
+                cb.on_close()
+
+            logging.info(f"{type(self).__name__} closed")
+            return jobs
 
         # Attempt to close tasks in loop
         if not self.loop.is_closed():
@@ -515,8 +512,7 @@ class Evaluator(abc.ABC):
 
             # Wait for tasks to be canceled
             if len(self._tasks_running) > 0:
-                self.loop.run_until_complete(self._await_cancelling_of_running_tasks())
-                self.process_local_tasks_done(self._tasks_done)
+                self.gather("ALL")
 
                 for job in self.jobs:
                     if job.status in [JobStatus.READY, JobStatus.RUNNING]:
@@ -526,8 +522,11 @@ class Evaluator(abc.ABC):
                             job.set_output("F_CANCELLED")
 
                         self._on_done(job)
-                        self.jobs_done.append(job)
                         self.job_id_gathered.append(job.id)
+                        jobs.append(job)
+
+                for cb in self._callbacks:
+                    cb.on_gather(jobs, [])
 
         self._tasks_done = []
         self._tasks_pending = []
@@ -536,6 +535,13 @@ class Evaluator(abc.ABC):
         if not self.loop.is_running():
             self.loop.close()
             self.loop = None
+
+        # calling callbacks
[…]
