### 2025-05-16 0060fea
chore (search): adding checkpoint_history_to_csv for all search init
