+++ tests/hpo/test__search_without_csv.py
@@ -14,8 +14,7 @@ def test_search_without_csv_dump():
     problem = HpProblem()
     problem.add_hyperparameter((0.0, 1.0), "x")
 
-    search = RandomSearch(problem, run)
-    search.history_to_csv = False
+    search = RandomSearch(problem, run, checkpoint_history_to_csv=False)
 
     max_evals = 100
     results = search.search(max_evals)
